% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/aquatic_interface.R
\name{add_oviposit}
\alias{add_oviposit}
\title{Add eggs from oviposition to aquatic model}
\usage{
add_oviposit(model, eggs)
}
\arguments{
\item{model}{an object from \link{make_MicroMoB}}

\item{eggs}{a vector of length \code{p} giving eggs for each place}
}
\description{
This method dispatches on the type of \code{model$aqua}
}
