% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/aquatic_trace.R
\name{add_oviposit.trace}
\alias{add_oviposit.trace}
\title{Add eggs from oviposition to forced aquatic model}
\usage{
\method{add_oviposit}{trace}(model, eggs)
}
\arguments{
\item{model}{an object from \link{make_MicroMoB}}

\item{eggs}{a vector of length \code{p} giving eggs for each place}
}
\description{
This function does nothing as trace models are not affected by
endogenous dynamics.
}
