% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/api_mosquito.R
\name{api_config_global}
\alias{api_config_global}
\title{Read global configuration options}
\usage{
api_config_global(path)
}
\arguments{
\item{path}{file path to a JSON file}
}
\description{
Read global configuration options
}
