% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mosquito_RM.R
\name{calculate_RM_equilibrium}
\alias{calculate_RM_equilibrium}
\title{Calculate approximate equilibrium for RM model}
\usage{
calculate_RM_equilibrium(model, X, H, b, c, r, f, q, g, EIP)
}
\arguments{
\item{model}{an object from \link{make_MicroMoB}}

\item{X}{a vector of infectious human population sizes}

\item{H}{a vector of total human population sizes}

\item{b}{transmission efficiency (mosquito to human)}

\item{c}{transmission efficiency (human to mosquito)}

\item{r}{recovery rate}

\item{f}{blood feeding rate}

\item{q}{human blood feeding fraction}

\item{g}{mortality rate}

\item{EIP}{length of extrinsic incubation period}
}
\description{
Using a differential equation representation, calculate an approximate
equilibrium for the RM model. This is only relevant for mosquito models
constructed with \link{setup_mosquito_RM}, aquatic models setup with
\link{setup_aqua_trace}, and human models setup with \link{setup_humans_SIS}.
This function only sets state variables in the mosquito object, and \code{lambda}
in the aqua object.
}
