% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ModelObject.R
\name{checkModObject}
\alias{checkModObject}
\title{Check a model object}
\usage{
checkModObject(Mod)
}
\arguments{
\item{Mod}{a model object (\link{list} or \link{environment})}
}
\description{
Check that a model object has minimum objects required to execute
a time step.
}
