% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Human.R
\name{compute.biteweight}
\alias{compute.biteweight}
\alias{compute.biteweight.null}
\title{Compute biting weight}
\usage{
compute.biteweight(biteweight, t)

\method{compute.biteweight}{null}(biteweight, t)
}
\arguments{
\item{biteweight}{an object from \link{setup.biteweight}}

\item{t}{time}
}
\description{
a thing
}
