% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Human.R
\name{compute.timespent.day}
\alias{compute.timespent.day}
\title{Compute human availability for daily time spent}
\usage{
\method{compute.timespent}{day}(tisp, biteweight, human, xi, t)
}
\arguments{
\item{tisp}{an object from \link{setup.timespent}}

\item{biteweight}{an object from \link{setup.biteweight}}

\item{human}{an object from \link{setup.human}}

\item{xi}{a vector of probabilities for each fraction of day (may be length 1)
giving the conditional probability of initiating a feeding search attempt at
that time given a feeding search occurs on that day.}

\item{t}{time}
}
\description{
Compute human availability for daily time spent
}
