% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Human.R
\name{compute.timespent}
\alias{compute.timespent}
\alias{compute.timespent.dt}
\title{#' @title Compute human availability (W)}
\usage{
compute.timespent(tisp, biteweight, human, xi, t)

\method{compute.timespent}{dt}(tisp, biteweight, human, xi, t)
}
\arguments{
\item{tisp}{an object from \link{setup.timespent}}

\item{biteweight}{an object from \link{setup.biteweight}}

\item{human}{an object from \link{setup.human}}

\item{xi}{a vector of probabilities of mosquito feeding initiation}

\item{t}{time}
}
\description{
do thing
}
\section{dispatching on a type of \code{timespent}}{
NA
}

