% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/other_hosts.R
\name{compute_B}
\alias{compute_B}
\title{Compute other host availability (\eqn{B})}
\usage{
compute_B(otherhosts, t)
}
\arguments{
\item{otherhosts}{an object from \link{setup_otherhosts}}

\item{t}{time}
}
\description{
Compute other host availability (\eqn{B})
}
\seealso{
\link{compute_B.simple}
}
