% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/other_hosts.R
\name{compute_B.simple}
\alias{compute_B.simple}
\title{Compute simple other host availability (\eqn{B})}
\usage{
\method{compute_B}{simple}(otherhosts, t)
}
\arguments{
\item{otherhosts}{an object from \link{setup_otherhosts}}

\item{t}{time}
}
\value{
a vector of dimension \eqn{p \times 1}
}
\description{
Compute simple other host availability (\eqn{B})
}
