% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/blood_feeding.R
\name{compute_beta.day}
\alias{compute_beta.day}
\title{Compute daily biting distribution matrix (\eqn{\beta})}
\usage{
\method{compute_beta}{day}(human, Psi_t, t)
}
\arguments{
\item{human}{an object from \link{setup_human}}

\item{Psi_t}{a time at risk matrix from \link{compute_Psi}}

\item{t}{time}
}
\value{
a matrix of dimension \eqn{n \times p}
}
\description{
Compute daily biting distribution matrix (\eqn{\beta})
}
