% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/blood_feeding.R
\name{compute_beta.dt}
\alias{compute_beta.dt}
\title{Compute fractional biting distribution array (\eqn{\beta})}
\usage{
\method{compute_beta}{dt}(human, Psi_t, t)
}
\arguments{
\item{human}{an object from \link{setup_human}}

\item{Psi_t}{a time at risk matrix from \link{compute_Psi}}

\item{t}{time}
}
\value{
an array of dimension \eqn{n \times p \times d}
}
\description{
Compute fractional biting distribution array (\eqn{\beta})
}
