% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/blood_feeding.R
\name{compute_beta}
\alias{compute_beta}
\title{Compute biting distribution matrix (\eqn{\beta})}
\usage{
compute_beta(human, Psi_t, t)
}
\arguments{
\item{human}{an object from \link{setup_human}}

\item{Psi_t}{a time at risk matrix from \link{compute_Psi}}

\item{t}{time}
}
\description{
Compute biting distribution matrix (\eqn{\beta})
}
\seealso{
\link{compute_beta.day} \link{compute_beta.dt}
}
