% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/blood_feeding.R
\name{compute_biting.day}
\alias{compute_biting.day}
\title{Compute parasite transmission terms from bloodmeals with daily human time spent (\eqn{\kappa}, \eqn{h})}
\usage{
\method{compute_biting}{day}(model, t)
}
\arguments{
\item{model}{a model object (an object from \link{setup_model_object})}

\item{t}{time}
}
\description{
Compute parasite transmission terms from bloodmeals with daily human time spent (\eqn{\kappa}, \eqn{h})
}
