% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/blood_feeding.R
\name{compute_biting.dt}
\alias{compute_biting.dt}
\title{Compute biting distribution and feeding rates with fractional human time spent (\eqn{\beta}, \eqn{fq})}
\usage{
\method{compute_biting}{dt}(model, t)
}
\arguments{
\item{model}{a model object (an object from \link{setup_model_object})}

\item{t}{time}
}
\description{
Compute biting distribution and feeding rates with fractional human time spent (\eqn{\beta}, \eqn{fq})
}
