% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/blood_feeding.R
\name{compute_biting}
\alias{compute_biting}
\title{Compute biting distribution and feeding rates (\eqn{\beta}, \eqn{fq})}
\usage{
compute_biting(model, t)
}
\arguments{
\item{model}{a model object (an object from \link{setup_model_object})}

\item{t}{time}
}
\description{
Compute biting distribution and feeding rates (\eqn{\beta}, \eqn{fq})
}
\seealso{
\link{compute_biting.dt} \link{compute_biting.day}
}
