% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/blood_feeding.R
\name{compute_bloodmeal.dt}
\alias{compute_bloodmeal.dt}
\title{Compute parasite transmission terms from bloodmeals with fractional human time spent (\eqn{\kappa}, \eqn{h})}
\usage{
\method{compute_bloodmeal}{dt}(model, t)
}
\arguments{
\item{model}{a model object (an \link{environment})}

\item{t}{time}
}
\description{
Compute parasite transmission terms from bloodmeals with fractional human time spent (\eqn{\kappa}, \eqn{h})
}
