% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/blood_feeding.R
\name{compute_bloodmeal}
\alias{compute_bloodmeal}
\title{Compute parasite transmission terms from bloodmeals (\eqn{\kappa}, \eqn{h})}
\usage{
compute_bloodmeal(model, t)
}
\arguments{
\item{model}{a model object (an \link{environment})}

\item{t}{time}
}
\description{
Compute parasite transmission terms from bloodmeals (\eqn{\kappa}, \eqn{h})
}
\seealso{
\link{compute_bloodmeal.dt} \link{compute_bloodmeal.day}
}
