% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bloodmeal.R
\name{compute_bloodmeal_simple}
\alias{compute_bloodmeal_simple}
\title{Compute bloodmeals taken by mosquitoes on hosts in simple models}
\usage{
compute_bloodmeal_simple(model)
}
\arguments{
\item{model}{an object from \link{make_MicroMoB}}
}
\value{
no return value
}
\description{
The difference between this and \link{compute_bloodmeal} is that
this function does not include any computations of alterative blood hosts
or visitors and is suitable for models which only include mosquitoes and resident
human populations.
}
