% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/aquatic_interface.R
\name{compute_emergents}
\alias{compute_emergents}
\title{Compute number of newly emerging adults (\eqn{\lambda})}
\usage{
compute_emergents(model)
}
\arguments{
\item{model}{an object from \link{make_MicroMoB}}
}
\description{
This method dispatches on the type of \code{model$aqua}
}
