% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/aquatic_trace.R
\name{compute_emergents.trace}
\alias{compute_emergents.trace}
\title{Compute number of newly emerging adults from forcing term}
\usage{
\method{compute_emergents}{trace}(model)
}
\arguments{
\item{model}{an object from \link{make_MicroMoB}}
}
\description{
This function dispatches on the second class attribute of \code{model$aqua}
for stochastic or deterministic behavior.
}
\details{
see \link{compute_emergents.trace_deterministic} and \link{compute_emergents.trace_stochastic}
}
