% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/aquatic_trace.R
\name{compute_emergents.trace_deterministic}
\alias{compute_emergents.trace_deterministic}
\title{Compute number of newly emerging adults from forcing term (deterministic)}
\usage{
\method{compute_emergents}{trace_deterministic}(model)
}
\arguments{
\item{model}{an object from \link{make_MicroMoB}}
}
\value{
a vector of length \code{l} giving the number of newly emerging adult in each patch
}
\description{
Return the column of the lambda matrix for this day.
}
