% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/aquatic_trace.R
\name{compute_emergents.trace_stochastic}
\alias{compute_emergents.trace_stochastic}
\title{Compute number of newly emerging adults from forcing term (stochastic)}
\usage{
\method{compute_emergents}{trace_stochastic}(model)
}
\arguments{
\item{model}{an object from \link{make_MicroMoB}}
}
\value{
a vector of length \code{l} giving the number of newly emerging adult in each patch
}
\description{
Draw a Poisson distributed number of emerging adults with mean parameter
from the column of the trace matrix for this day.
}
