% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mosquito_trace.R
\name{compute_f.trace}
\alias{compute_f.trace}
\title{Compute null mosquito feeding rate (\eqn{f})}
\usage{
\method{compute_f}{trace}(model, B)
}
\arguments{
\item{model}{an object from \link{make_MicroMoB}}

\item{B}{a vector of length \code{p} giving total blood host availability by patch}
}
\description{
Compute null mosquito feeding rate (\eqn{f})
}
