% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/humans_MOI.R
\name{compute_H.MOI}
\alias{compute_H.MOI}
\title{Compute human population strata sizes for MOI model (\eqn{H})}
\usage{
\method{compute_H}{MOI}(model)
}
\arguments{
\item{model}{an object from \link{make_MicroMoB}}
}
\description{
Compute human population strata sizes for MOI model (\eqn{H})
}
