% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pathogen.R
\name{compute_h}
\alias{compute_h}
\title{Compute force of infection on human hosts (\eqn{h})}
\usage{
compute_h(model, t)
}
\arguments{
\item{model}{a model object (an object from \link{setup_model_object})}

\item{t}{time}
}
\description{
\eqn{h}, the force of infection on each human strata, is calculated
after calling \link{compute_biting} which computes the biting distribution matrix
and feeding rates.
}
