% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/humans_SIP.R
\name{compute_H.SIP}
\alias{compute_H.SIP}
\title{Compute human population strata sizes for SIP model (\eqn{H})}
\usage{
\method{compute_H}{SIP}(model)
}
\arguments{
\item{model}{an object from \link{make_MicroMoB}}
}
\value{
a vector of length \code{n} giving the size of each human population stratum
}
\description{
Compute human population strata sizes for SIP model (\eqn{H})
}
