% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/humans_SIR.R
\name{compute_H.SIR}
\alias{compute_H.SIR}
\title{Compute human population strata sizes for SIR model (\eqn{H})}
\usage{
\method{compute_H}{SIR}(model)
}
\arguments{
\item{model}{an object from \link{make_MicroMoB}}
}
\description{
Compute human population strata sizes for SIR model (\eqn{H})
}
