% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/humans_SIS.R
\name{compute_H.SIS}
\alias{compute_H.SIS}
\title{Compute human population strata sizes for SIS model (\eqn{H})}
\usage{
\method{compute_H}{SIS}(model)
}
\arguments{
\item{model}{an object from \link{make_MicroMoB}}
}
\description{
Compute human population strata sizes for SIS model (\eqn{H})
}
