% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/other_interface.R
\name{compute_O}
\alias{compute_O}
\title{Compute available alternative blood hosts (\eqn{O})}
\usage{
compute_O(model)
}
\arguments{
\item{model}{an object from \link{make_MicroMoB}}
}
\value{
a vector of length \code{p}
}
\description{
This method dispatches on the type of \code{model$alternative}.
}
