% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/other_trace.R
\name{compute_O.trace}
\alias{compute_O.trace}
\title{Compute available alternative blood hosts for trace model (\eqn{O})}
\usage{
\method{compute_O}{trace}(model)
}
\arguments{
\item{model}{an object from \link{make_MicroMoB}}
}
\value{
a vector of length \code{p} giving biting availability of other blood hosts at each patch
}
\description{
Compute available alternative blood hosts for trace model (\eqn{O})
}
