% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mosquito_BQ.R
\name{compute_oviposit.BQ}
\alias{compute_oviposit.BQ}
\title{Compute number of eggs laid from oviposition for each aquatic habitat for BQ model}
\usage{
\method{compute_oviposit}{BQ}(model)
}
\arguments{
\item{model}{an object from \link{make_MicroMoB}}
}
\value{
a vector of length \code{l} giving the total number of eggs laid by adult mosquitoes in each aquatic habitat
}
\description{
This method returns a vector of length \code{l}.
}
\details{
see \link{compute_oviposit.BQ_deterministic} and \link{compute_oviposit.BQ_stochastic}
}
