% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mosquito_BQ.R
\name{compute_oviposit.BQ_deterministic}
\alias{compute_oviposit.BQ_deterministic}
\title{Compute number of eggs laid from oviposition for each patch for deterministic RM model}
\usage{
\method{compute_oviposit}{BQ_deterministic}(model)
}
\arguments{
\item{model}{an object from \link{make_MicroMoB}}
}
\value{
a vector of length \code{l} giving the total number of eggs laid by adult mosquitoes in each aquatic habitat
}
\description{
Compute number of eggs laid from oviposition for each patch for deterministic RM model
}
