% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mosquito_RM.R
\name{compute_oviposit.RM_deterministic}
\alias{compute_oviposit.RM_deterministic}
\title{Compute number of eggs laid from oviposition for each patch for deterministic RM model}
\usage{
\method{compute_oviposit}{RM_deterministic}(model)
}
\arguments{
\item{model}{an object from \link{make_MicroMoB}}
}
\description{
Compute number of eggs laid from oviposition for each patch for deterministic RM model
}
