% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mosquito_RM.R
\name{compute_oviposit.RM_stochastic}
\alias{compute_oviposit.RM_stochastic}
\title{Compute number of eggs laid from oviposition for each patch for stochastic RM model}
\usage{
\method{compute_oviposit}{RM_stochastic}(model)
}
\arguments{
\item{model}{an object from \link{make_MicroMoB}}
}
\value{
a vector of length \code{p} giving the total number of eggs laid by adult mosquitoes in each patch
}
\description{
Compute number of eggs laid from oviposition for each patch for stochastic RM model
}
