% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mosquito_trace.R
\name{compute_oviposit.trace}
\alias{compute_oviposit.trace}
\title{Compute number of eggs laid from oviposition for each patch for null model}
\usage{
\method{compute_oviposit}{trace}(model)
}
\arguments{
\item{model}{an object from \link{make_MicroMoB}}
}
\value{
a vector of length \code{p} giving the total number of eggs laid by adult mosquitoes in each patch
}
\description{
This method dispatches on the type of \code{model$mosquito}
}
