% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/humans_interface.R
\name{compute_Psi}
\alias{compute_Psi}
\title{Compute time at risk matrix (\eqn{\Psi})}
\usage{
compute_Psi(model)
}
\arguments{
\item{model}{an object from \link{make_MicroMoB}}
}
\value{
a matrix with \code{n} rows and \code{p} columns
}
\description{
The time at risk matrix is \eqn{\Psi = \Theta \xi} This method dispatches on the type of \code{model$human}.
}
