% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/humans_SIS.R
\name{compute_Psi.SIS}
\alias{compute_Psi.SIS}
\title{Compute time at risk matrix for SIS model (\eqn{\Psi})}
\usage{
\method{compute_Psi}{SIS}(model)
}
\arguments{
\item{model}{an object from \link{make_MicroMoB}}
}
\description{
Compute time at risk matrix for SIS model (\eqn{\Psi})
}
