% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/time_spent.R
\name{compute_Psi.timespent.day}
\alias{compute_Psi.timespent.day}
\title{Compute time at risk for daily time spent (\eqn{\Psi^{T}})}
\usage{
\method{compute_Psi.timespent}{day}(human, xi, t)
}
\arguments{
\item{human}{an object from \link{setup.human}}

\item{xi}{probabilities to initiate a feeding search during each fraction
of the day (must sum to 1)}

\item{t}{time (day)}
}
\value{
a matrix of dimension \eqn{p \times n}
}
\description{
Compute time at risk for daily time spent (\eqn{\Psi^{T}})
}
