% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/time_spent.R
\name{compute_Psi.timespent.dt}
\alias{compute_Psi.timespent.dt}
\title{Compute time at risk for fractional daily time spent (\eqn{\Psi^{T}})}
\usage{
\method{compute_Psi.timespent}{dt}(tisp, xi, t)
}
\arguments{
\item{tisp}{a time spent object (see \link{setup.timespent})}

\item{xi}{probabilities to initiate a feeding search during each fraction
of the day (must sum to 1)}

\item{t}{time (day)}
}
\description{
Compute time at risk for fractional daily time spent (\eqn{\Psi^{T}})
}
