% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/time_spent.R
\name{compute_Psi.timespent}
\alias{compute_Psi.timespent}
\title{Compute time at risk (\eqn{\Psi})}
\usage{
compute_Psi.timespent(human, xi, t)
}
\arguments{
\item{human}{an object from \link{setup.human}}

\item{xi}{probabilities to initiate a feeding search during each fraction
of the day (must sum to 1)}

\item{t}{time (day)}
}
\value{
a (transposed) time at risk \link{matrix} or \link{array}, which should be attached to the \code{tisp}
object and stored for later computation (human availability and the biting distribution matrix)
}
\description{
Compute time at risk (\eqn{\Psi})
}
\seealso{
\link{compute_Psi.timespent.day} \link{compute_Psi.timespent.dt}
}
