% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mosquito_BQ.R
\name{compute_q.BQ}
\alias{compute_q.BQ}
\title{Compute human blood feeding fraction for BQ model (\eqn{q})}
\usage{
\method{compute_q}{BQ}(model, W, Wd, B)
}
\arguments{
\item{model}{an object from \link{make_MicroMoB}}

\item{W}{a vector of length \code{p} giving human availability by patch (\eqn{W})}

\item{Wd}{a vector of length \code{p} giving visitor availability by patch (\eqn{W_{\delta}})}

\item{B}{a vector of length \code{p} giving total blood host availability by patch (\eqn{B})}
}
\value{
a vector of length \code{p} giving the proportion of bites taken on human hosts in each blood feeding haunt
}
\description{
The human blood feeding fraction is simply the proportion of human hosts.
}
