% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mosquito_interface.R
\name{compute_q}
\alias{compute_q}
\title{Compute human blood feeding fraction (\eqn{q})}
\usage{
compute_q(model, W, Wd, B)
}
\arguments{
\item{model}{an object from \link{make_MicroMoB}}

\item{W}{a vector of length \code{p} giving human availability by patch (\eqn{W})}

\item{Wd}{a vector of length \code{p} giving visitor availability by patch (\eqn{W_{\delta}})}

\item{B}{a vector of length \code{p} giving total blood host availability by patch (\eqn{B})}
}
\value{
a vector of length \code{p}
}
\description{
This method dispatches on the type of \code{model$mosquito}
}
