% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mosquito_RM.R
\name{compute_q.RM}
\alias{compute_q.RM}
\title{Compute human blood feeding fraction for RM model (\eqn{q})}
\usage{
\method{compute_q}{RM}(model, W, Wd, B)
}
\arguments{
\item{model}{an object from \link{make_MicroMoB}}

\item{W}{a vector of length \code{p} giving human availability by patch (\eqn{W})}

\item{Wd}{a vector of length \code{p} giving visitor availability by patch (\eqn{W_{\delta}})}

\item{B}{a vector of length \code{p} giving total blood host availability by patch (\eqn{B})}
}
\value{
a vector of length \code{p} giving the proportion of bites taken on human hosts in each patch
}
\description{
This method simply returns the \code{q} parameter of the mosquito object,
because the RM model assumes a constant fraction of blood meals are taken on
human hosts.
}
