% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mosquito_trace.R
\name{compute_q.trace}
\alias{compute_q.trace}
\title{Compute null human blood feeding fraction (\eqn{q})}
\usage{
\method{compute_q}{trace}(model, W, Wd, B)
}
\arguments{
\item{model}{an object from \link{make_MicroMoB}}

\item{W}{a vector of length \code{p} giving human availability by patch (\eqn{W})}

\item{Wd}{a vector of length \code{p} giving visitor availability by patch (\eqn{W_{\delta}})}

\item{B}{a vector of length \code{p} giving total blood host availability by patch (\eqn{B})}
}
\description{
Compute null human blood feeding fraction (\eqn{q})
}
