% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/compute_terms.R
\name{compute_terms}
\alias{compute_terms}
\title{Compute dynamical terms}
\usage{
compute_terms(varslist, dts_out, pars, s, i)
}
\arguments{
\item{varslist}{a \link{list} with variable names attached}

\item{dts_out}{a \link{matrix}, the output of deSolve}

\item{pars}{a \link{list}}

\item{s}{the vector species index}

\item{i}{the host species index}
}
\value{
\link{matrix}
}
\description{
Using the output of deSolve compute the dynamical terms every point in time. This method dispatches on the type of \code{pars$compute}.
}
