% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/time_spent.R
\name{compute_W.dt}
\alias{compute_W.dt}
\title{Compute human availability for fractional daily time spent (W)}
\usage{
\method{compute_W}{dt}(human, t)
}
\arguments{
\item{human}{an object from \link{setup_human}}

\item{t}{time}
}
\value{
a matrix of dimension \eqn{p \times d}
}
\description{
For each time period the human availability is \eqn{W\[t\] = \Psi\[t\]^{T} \cdot w_f H},
and the sum of them over the day is the overall human availability.
}
