% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/humans_interface.R
\name{compute_W}
\alias{compute_W}
\title{Compute available humans (\eqn{W})}
\usage{
compute_W(model)
}
\arguments{
\item{model}{an object from \link{make_MicroMoB}}
}
\value{
a vector of length \code{p}
}
\description{
This is normally computed as \deqn{W = \Psi^{\intercal} \cdot w_{f} H}.
This method dispatches on the type of \code{model$human}.
}
