% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/humans_SIS.R
\name{compute_W.SIS}
\alias{compute_W.SIS}
\title{Compute available humans for SIS model}
\usage{
\method{compute_W}{SIS}(model)
}
\arguments{
\item{model}{an object from \link{make_MicroMoB}}
}
\description{
Compute available humans for SIS model
}
