% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/time_spent.R
\name{compute_W.timespent.day}
\alias{compute_W.timespent.day}
\title{Compute human availability for daily time spent (W)}
\usage{
\method{compute_W.timespent}{day}(human, t)
}
\arguments{
\item{human}{an object from \link{setup.human}}

\item{t}{time}
}
\value{
a vector of length \eqn{p}
}
\description{
The human availability is \eqn{\eqn{W = \Psi^{T} \cdot w_f H},}
}
