% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/time_spent.R
\name{compute_W.timespent.dt}
\alias{compute_W.timespent.dt}
\title{Compute human availability for fractional daily time spent (W)}
\usage{
\method{compute_W.timespent}{dt}(tisp, biteweight, human, t)
}
\arguments{
\item{tisp}{an object from \link{setup.timespent}}

\item{biteweight}{an object from \link{setup.biteweight}}

\item{human}{an object from \link{setup.human}}

\item{t}{time}
}
\description{
For each time period the human availability is \eqn{W\[t\] = \Psi\[t\]^{T} \cdot w_f H},
and the sum of them over the day is the overall human availability.
}
