% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/time_spent.R
\name{compute_W.timespent}
\alias{compute_W.timespent}
\title{Compute human availability (W)}
\usage{
compute_W.timespent(human, t)
}
\arguments{
\item{human}{an object from \link{setup.human}}

\item{t}{time}
}
\description{
Compute human availability (W)
}
\seealso{
\link{compute_W.timespent.day} \link{compute_W.timespent.dt}
}
