% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/other_hosts.R
\name{compute_W_delta}
\alias{compute_W_delta}
\title{Compute visitor availability (\eqn{W_{\delta}})}
\usage{
compute_W_delta(visitors, t)
}
\arguments{
\item{visitors}{an object from \link{setup_visitors}}

\item{t}{time}
}
\description{
Compute visitor availability (\eqn{W_{\delta}})
}
\seealso{
\link{compute_W_delta.simple}
}
