% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/other_hosts.R
\name{compute_W_delta.simple}
\alias{compute_W_delta.simple}
\title{Compute simple visitor availability (\eqn{W_{\delta}})}
\usage{
\method{compute_W_delta}{simple}(visitors, t)
}
\arguments{
\item{visitors}{an object from \link{setup_visitors}}

\item{t}{time}
}
\value{
a vector of dimension \eqn{p \times 1}
}
\description{
Compute simple visitor availability (\eqn{W_{\delta}})
}
