% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/visitor_interface.R
\name{compute_Wd}
\alias{compute_Wd}
\title{Compute available visitors (\eqn{W_{\delta}})}
\usage{
compute_Wd(model)
}
\arguments{
\item{model}{an object from \link{make_MicroMoB}}
}
\value{
a vector of length \code{p} giving biting availability of visitors at each patch
}
\description{
This method dispatches on the type of \code{model$visitor}.
}
