% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/visitor_trace.R
\name{compute_Wd.trace}
\alias{compute_Wd.trace}
\title{Compute available visitors for trace model (\eqn{W_{\delta}})}
\usage{
\method{compute_Wd}{trace}(model)
}
\arguments{
\item{model}{an object from \link{make_MicroMoB}}
}
\description{
Compute available visitors for trace model (\eqn{W_{\delta}})
}
