% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/human.R
\name{compute_wf.biteweight}
\alias{compute_wf.biteweight}
\title{Compute biting weight}
\usage{
compute_wf.biteweight(biteweight, t)
}
\arguments{
\item{biteweight}{an object from \link{setup.biteweight}}

\item{t}{time}
}
\description{
Return the biting weights for humans at a given time
}
