% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/human.R
\name{compute_wf.biteweight.simple}
\alias{compute_wf.biteweight.simple}
\title{Compute biting weight for simple model}
\usage{
\method{compute_wf.biteweight}{simple}(biteweight, t)
}
\arguments{
\item{biteweight}{an object from \link{setup.biteweight}}

\item{t}{time}
}
\description{
Compute biting weight for simple model
}
