% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/humans_MOI.R
\name{compute_wf.MOI}
\alias{compute_wf.MOI}
\title{Compute human biting weights for MOI model (\eqn{w_{f}})}
\usage{
\method{compute_wf}{MOI}(model)
}
\arguments{
\item{model}{an object from \link{make_MicroMoB}}
}
\description{
Compute human biting weights for MOI model (\eqn{w_{f}})
}
