% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/human.R
\name{compute_wf}
\alias{compute_wf}
\title{Compute biting weight}
\usage{
compute_wf(biteweight, t)
}
\arguments{
\item{biteweight}{an object from \link{setup_biteweight}}

\item{t}{time}
}
\description{
Return the biting weights for humans at a given time
}
\seealso{
\link{compute_wf.simple}
}
