% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/humans_SIS.R
\name{compute_wf.SIS}
\alias{compute_wf.SIS}
\title{Compute human biting weights for SIS model (\eqn{w_{f}})}
\usage{
\method{compute_wf}{SIS}(model)
}
\arguments{
\item{model}{an object from \link{make_MicroMoB}}
}
\description{
Compute human biting weights for SIS model (\eqn{w_{f}})
}
