% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pathogen.R
\name{compute_x}
\alias{compute_x}
\title{Compute resident net infectiousness (\eqn{x})}
\usage{
compute_x(model, t)
}
\arguments{
\item{model}{an object from \link{setup_model_object}}

\item{t}{time}
}
\description{
\eqn{x}, the net infectiousness of residents, is the probability
that a mosquito would become infected after biting a resident It is the prevalence
multiplied by the transmission efficiency.
}
\seealso{
\link{compute_x.rm}
}
