% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pathogen.R
\name{compute_x.rm}
\alias{compute_x.rm}
\title{Compute simple visitor availability (\eqn{x})}
\usage{
\method{compute_x}{rm}(pathogen, t)
}
\arguments{
\item{pathogen}{an object from \link{setup_pathogen}}

\item{t}{time}
}
\value{
a vector of dimension \eqn{p \times 1}
}
\description{
Compute simple visitor availability (\eqn{x})
}
