% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/humans_SIR.R
\name{compute_x.SIR}
\alias{compute_x.SIR}
\title{Compute net infectiousness for SIR model (\eqn{x})}
\usage{
\method{compute_x}{SIR}(model)
}
\arguments{
\item{model}{an object from \link{make_MicroMoB}}
}
\value{
a vector of length \code{n} giving the net infectiousness of human hosts in each stratum
}
\description{
Compute net infectiousness for SIR model (\eqn{x})
}
