% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/visitor_interface.R
\name{compute_xd}
\alias{compute_xd}
\title{Compute net infectiousness of visitors (\eqn{x_{\delta}})}
\usage{
compute_xd(model)
}
\arguments{
\item{model}{an object from \link{make_MicroMoB}}
}
\value{
a vector of length \code{p} giving net infectiousness of visitors at each patch
}
\description{
This method dispatches on the type of \code{model$visitor}.
}
