% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mosquito_BQ.R
\name{compute_Z.BQ}
\alias{compute_Z.BQ}
\title{Compute density of infective mosquitoes for BQ model (\eqn{Z})}
\usage{
\method{compute_Z}{BQ}(model)
}
\arguments{
\item{model}{an object from \link{make_MicroMoB}}
}
\value{
a vector of length \code{p} giving the density of infected and infectious mosquitoes in each blood feeding haunt
}
\description{
This method returns \code{Z}.
}
