% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mosquito_interface.R
\name{compute_Z}
\alias{compute_Z}
\title{Compute density of infective mosquitoes (\eqn{Z})}
\usage{
compute_Z(model)
}
\arguments{
\item{model}{an object from \link{make_MicroMoB}}
}
\value{
a vector of length \code{p}
}
\description{
This method dispatches on the type of \code{model$mosquito}
}
\details{
#' @title Compute human blood feeding fraction (\eqn{q})
#' @description This method dispatches on the type of \code{model$mosquito}
#' @param model an object from \link{make_MicroMoB}
#' @return a vector of length \code{p}
#' @export
compute_q <- function(model) {
UseMethod("compute_q", model$mosquito)
}
}
