% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mosquito.R
\name{compute_Z.trace}
\alias{compute_Z.trace}
\title{Compute simple visitor availability (\eqn{W_{\delta}})}
\usage{
\method{compute_Z}{trace}(mosquito, t)
}
\arguments{
\item{mosquito}{an object from \link{setup_mosquito}}

\item{t}{time}
}
\value{
a vector of dimension \eqn{p \times 1}
}
\description{
Compute simple visitor availability (\eqn{W_{\delta}})
}
