% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils_internal.R
\name{distribute}
\alias{distribute}
\title{Distribute items into bins as evenly as possible}
\usage{
distribute(n, p)
}
\arguments{
\item{n}{number of bins}

\item{p}{number of itens}
}
\value{
a numeric vector of bin sizes
}
\description{
Distribute items into bins as evenly as possible
}
