% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils_internal.R
\name{divmod}
\alias{divmod}
\title{Division of integers}
\usage{
divmod(a, b)
}
\arguments{
\item{a}{the dividend}

\item{b}{the divisor}
}
\value{
a list with two elements, \code{quo} (quotient) and \code{rem} (remainder)
}
\description{
Division of integers
}
