% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{draw_multinom}
\alias{draw_multinom}
\title{Draw a multinomially distributed random vector}
\usage{
draw_multinom(n, prob)
}
\arguments{
\item{n}{an integer giving the number of balls to distribute in bins}

\item{prob}{a vector of probabilities for each bin, which must sum to one}
}
\value{
an integer vector of length equal to the length of \code{prob}
}
\description{
Warning: this function does no argument checking. Ensure the arguments
are as follows.
}
\note{
This function uses the algorithm presented in:
Startek, Michał. "An asymptotically optimal, online algorithm for weighted random sampling with replacement." arXiv preprint arXiv:1611.00532 (2016).
}
