% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot-MYZ.R
\name{dts_lines_M}
\alias{dts_lines_M}
\title{Add lines for adult mosquito population density}
\usage{
dts_lines_M(MYZ, pars, clrs = "darkblue", llty = 1)
}
\arguments{
\item{MYZ}{a list the output of \code{ramp.dts::parse_deout()}}

\item{pars}{a list that defines an \code{ramp.dts} model (\emph{e.g.},  generated by \code{dts_setup()})}

\item{clrs}{a vector of colors}

\item{llty}{an integer (or integers) to set the \code{lty} for plotting}
}
\description{
Add lines for adult mosquito population density
}
