% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/human-SIP.R
\name{dts_lines_X_SIP}
\alias{dts_lines_X_SIP}
\title{Add lines for the density of infected individuals for the SIP model}
\usage{
dts_lines_X_SIP(
  XH,
  pars,
  clrs = c("darkblue", "darkred", "darkgreen"),
  llty = 1
)
}
\arguments{
\item{XH}{a list with the outputs of parse_dts_out_X.SIP}

\item{pars}{a list that defines an \code{ramp.dts} model (\emph{e.g.},  generated by \code{dts_setup()})}

\item{clrs}{a vector of colors}

\item{llty}{an integer (or integers) to set the \code{lty} for plotting}
}
\description{
Add lines for the density of infected individuals for the SIP model
}
