% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/human-SIS.R
\name{dts_lines_X_SIS}
\alias{dts_lines_X_SIS}
\title{Add lines for the density of infected individuals for the SIS model}
\usage{
dts_lines_X_SIS(XH, nStrata, clrs = c("darkblue", "darkred"), llty = 1)
}
\arguments{
\item{XH}{a list with the outputs of parse_dts_out_X_SIS}

\item{nStrata}{the number of population strata}

\item{clrs}{a vector of colors}

\item{llty}{an integer (or integers) to set the \code{lty} for plotting}
}
\description{
Add lines for the density of infected individuals for the SIS model
}
