% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot-MYZ.R
\name{dts_lines_YZ_fracs}
\alias{dts_lines_YZ_fracs}
\title{Add lines for the fraction of infected and infective mosquitoes}
\usage{
dts_lines_YZ_fracs(MYZ, pars, Yclrs = "purple", Zclrs = "darkred", llty = 1)
}
\arguments{
\item{MYZ}{a list the output of \code{ramp.dts::parse_deout()}}

\item{pars}{a list that defines an \code{ramp.dts} model (\emph{e.g.},  generated by \code{dts_setup()})}

\item{Yclrs}{a vector of colors for infected mosquitoes}

\item{Zclrs}{a vector of colors for infective mosquitoes}

\item{llty}{an integer (or integers) to set the \code{lty} for plotting}
}
\description{
Add lines for the fraction of infected and infective mosquitoes
}
