% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot-MYZ.R
\name{dts_plot_M}
\alias{dts_plot_M}
\title{Plot adult mosquito population density}
\usage{
dts_plot_M(
  pars,
  s = 1,
  clrs = "darkblue",
  llty = 1,
  stable = FALSE,
  add_axes = TRUE
)
}
\arguments{
\item{pars}{a list that defines an \code{ramp.dts} model (\emph{e.g.},  generated by \code{dts_setup()})}

\item{s}{the vector species index}

\item{clrs}{a vector of colors}

\item{llty}{an integer (or integers) to set the \code{lty} for plotting}

\item{stable}{a logical: set to FALSE for \code{orbits} and TRUE for \code{stable_orbits}}

\item{add_axes}{a logical: plot axes only if TRUE}
}
\description{
Plot adult mosquito population density
}
