% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot-terms.R
\name{dts_plot_PR}
\alias{dts_plot_PR}
\title{Plot the prevalence / parasite rate (PR) from a model of human infection and immunity}
\usage{
dts_plot_PR(
  pars,
  i = 1,
  clrs = "black",
  llty = 1,
  stable = FALSE,
  add_axes = TRUE
)
}
\arguments{
\item{pars}{a list that defines an \code{ramp.dts} model (\emph{e.g.},  generated by \code{dts_setup()})}

\item{i}{the host species index}

\item{clrs}{a vector of colors}

\item{llty}{an integer (or integers) that specifies \code{lty} for plotting}

\item{stable}{a logical: set to FALSE for \code{orbits} and TRUE for \code{stable_orbits}}

\item{add_axes}{a logical: plot axes only if TRUE}
}
\description{
Plot the prevalence / parasite rate (PR) from a model of human infection and immunity
}
