% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/human-interface.R
\name{dts_plot_X}
\alias{dts_plot_X}
\title{Basic plotting for epidemiological models}
\usage{
dts_plot_X(
  pars,
  i = 1,
  clrs = "black",
  llty = 1,
  stable = FALSE,
  add_axes = TRUE
)
}
\arguments{
\item{pars}{a list that defines an \code{ramp.dts} model (\emph{e.g.},  generated by \code{dts_setup()})}

\item{i}{the host species index}

\item{clrs}{a vector of colors}

\item{llty}{an integer (or integers) to set the \code{lty} for plotting}

\item{stable}{a logical: set to FALSE for \code{orbits} and TRUE for \code{stable_orbits}}

\item{add_axes}{a logical: plot axes only if TRUE}
}
\description{
Basic plotting for epidemiological models
}
