% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot-MYZ.R
\name{dts_plot_YZ_fracs}
\alias{dts_plot_YZ_fracs}
\title{Plot the fraction of infected and infective mosquitoes}
\usage{
dts_plot_YZ_fracs(
  pars,
  s = 1,
  Yclrs = "purple",
  Zclrs = "darkred",
  llty = 1,
  stable = FALSE,
  add_axes = TRUE
)
}
\arguments{
\item{pars}{a list that defines an \code{ramp.dts} model (\emph{e.g.},  generated by \code{dts_setup()})}

\item{s}{the vector species index}

\item{Yclrs}{a vector of colors for infected mosquitoes}

\item{Zclrs}{a vector of colors for infective mosquitoes}

\item{llty}{an integer (or integers) to set the \code{lty} for plotting}

\item{stable}{a logical, set to FALSE for \code{orbits} and TRUE for \code{stable_orbits}}

\item{add_axes}{a logical to plot axes only if TRUE}
}
\description{
Plot the fraction of infected and infective mosquitoes
}
