% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dts_solve.R
\name{dts_solve.dts}
\alias{dts_solve.dts}
\title{Solve a system of equations as an ode}
\usage{
\method{dts_solve}{dts}(pars, Tmax = 365)
}
\arguments{
\item{pars}{a \link{list} that defines a model}

\item{Tmax}{the last time point, run from 0...Tmax}
}
\value{
a \link{list}
}
\description{
Implements \link{dts_solve} for ordinary differential equations
}
