% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dts_solve.R
\name{dts_solve.mosy}
\alias{dts_solve.mosy}
\title{Solve a system of equations for mosquito ecology using dts_diffeqn_mosy}
\usage{
\method{dts_solve}{mosy}(pars, Tmax = 365)
}
\arguments{
\item{pars}{a \link{list} that defines a model}

\item{Tmax}{the last time point, run from 0...Tmax}
}
\value{
a \link{list}
}
\description{
Implements \link{dts_solve} for mosquito dynamics (no transmission)
}
