% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dts_solve.R
\name{dts_stable_orbit}
\alias{dts_stable_orbit}
\title{Solve for the steady state or stable orbit of a system of equations}
\usage{
dts_stable_orbit(pars, Tburn = 10, yr = 365)
}
\arguments{
\item{pars}{a \link{list} that defines a model}

\item{Tburn}{the number of steps to burn}

\item{yr}{the number of time steps in a year}
}
\value{
a \link{list}
}
\description{
This method dispatches on the type of \code{pars$dts}.
}
