% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dts_equations.R
\name{DTS_step}
\alias{DTS_step}
\title{Generalized spatial differential equation model}
\usage{
DTS_step(t, y, pars)
}
\arguments{
\item{t}{current simulation time}

\item{y}{state vector}

\item{pars}{a \link{list}}
}
\value{
a \link{list} containing the vector of all state derivatives
}
\description{
Compute and update the state variables for
the generic model
}
