% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dts_equations.R
\name{DTS_step_cohort}
\alias{DTS_step_cohort}
\title{Difference equation models for human cohorts}
\usage{
DTS_step_cohort(a, y, pars, F_eir)
}
\arguments{
\item{a}{age of a cohort}

\item{y}{state vector}

\item{pars}{a \link{list}}

\item{F_eir}{a trace function that returns the eir as a function of time}
}
\value{
a \link{vector} containing the vector of all state derivatives
}
\description{
Compute and update the state variables for
a cohort
}
