% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dts_equations.R
\name{DTS_step_human}
\alias{DTS_step_human}
\title{Difference equations isolating the humans, forced with Ztrace}
\usage{
DTS_step_human(t, y, pars)
}
\arguments{
\item{t}{current simulation time}

\item{y}{state vector}

\item{pars}{a \link{list}}
}
\value{
a \link{vector} containing the vector of all state derivatives
}
\description{
Compute and update the state variables for
a model with only humans
}
