% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dts_equations.R
\name{DTS_step_mosy}
\alias{DTS_step_mosy}
\title{Generalized spatial differential equation model (mosquito only)}
\usage{
DTS_step_mosy(t, y, pars)
}
\arguments{
\item{t}{current simulation time}

\item{y}{state vector}

\item{pars}{a \link{list}
the appropriate adult mosquito model}
}
\value{
a \link{vector} containing the vector of all state derivatives
}
\description{
Compute and update the state variables for
a model with mosquito ecology (no transmission)
}
