% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/eip_fixed.R
\name{EIP.fixed}
\alias{EIP.fixed}
\title{Modify parameters due to exogenous forcing by all kinds of control}
\usage{
\method{EIP}{fixed}(t, EIPmod)
}
\arguments{
\item{t}{current simulation time}

\item{EIPmod}{a \link{list}}
}
\value{
\link{numeric}
}
\description{
Implements \link{EIP} for the fixed model (the EIP is constant)
}
