% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/eip_null.R
\name{EIP.null}
\alias{EIP.null}
\title{Modify parameters due to exogenous forcing by all kinds of control}
\usage{
\method{EIP}{null}(t, MYZpar)
}
\arguments{
\item{t}{current simulation time}

\item{MYZpar}{a \link{list}}
}
\value{
\link{numeric}
}
\description{
Implements \link{EIP} for the null model
}
