% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/adult-bionomics.R
\name{F_f.dddn}
\alias{F_f.dddn}
\title{Dawn, day, dusk, night model for the blood feeding rate}
\usage{
\method{F_f}{dddn}(t, MYZpar)
}
\arguments{
\item{t}{current simulation time}

\item{MYZpar}{a \link{list}}
}
\value{
a \link{numeric} vector of length \code{nPatches}
}
\description{
Implements \link{F_f} for a dddn model
}
