% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{Ff_0}
\alias{Ff_0}
\title{Functional form type 0}
\usage{
Ff_0(x, fx = 1, sf = 1)
}
\arguments{
\item{x}{the value}

\item{fx}{the asymptote}

\item{sf}{controls slope}
}
\description{
An equation of the form
\deqn{f(x) = \frac{f_{x} s_{f} x}{1 + s_{f} x}}
which has an asymptote at \eqn{f_{x}} and \eqn{s_{f}} controls the rate at which
the asymptote is reached.
}
