% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/other_trace.R
\name{get_config_alternative_trace}
\alias{get_config_alternative_trace}
\title{Get parameters for trace driven alternative blood hosts}
\usage{
get_config_alternative_trace(path)
}
\arguments{
\item{path}{a file path to a JSON file}
}
\value{
a named \link{list}
}
\description{
The JSON config file should have two entries:
\itemize{
\item O: vector or matrix (see \link{time_patch_varying_parameter} for valid dimensions)
}

For interpretation of the entries, please read \link{setup_alternative_trace}.
}
\examples{
# to see an example of proper JSON input, run the following
library(jsonlite)
par <- list(
 "O" = rep(1, 5)
)
toJSON(par)
}
