% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/config.R
\name{get_config_mosquito_MicroMoB}
\alias{get_config_mosquito_MicroMoB}
\title{Get parameters for configuring a mosquito only Micro-MoB simulation}
\usage{
get_config_mosquito_MicroMoB(path)
}
\arguments{
\item{path}{a file path to a JSON file}
}
\value{
a named \link{list}
}
\description{
The JSON config file should have:
\itemize{
\item tmax: integer
\item p: integer
\item aqua_path: file path
\item aqua_model: character in \code{BH}, \code{trace}
\item adult_path: file path
\item adult_model: character in \code{RM}
}
}
