% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mosquito_trace.R
\name{get_config_mosquito_trace}
\alias{get_config_mosquito_trace}
\title{Get parameters for null mosquito model}
\usage{
get_config_mosquito_trace(path)
}
\arguments{
\item{path}{a file path to a JSON file}
}
\value{
a named \link{list}
}
\description{
The JSON config file should have 1 entry:
\itemize{
\item oviposit: vector
}

For interpretation of the entries, please read \link{setup_mosquito_trace}.
}
\examples{
# to see an example of proper JSON input, run the following
library(jsonlite)
par <- list(
 "oviposit" = rep(1, 5)
)
toJSON(par, pretty = TRUE)
}
