% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mosquito_RM_API.R
\name{get_eip_mosquito_RM}
\alias{get_eip_mosquito_RM}
\title{Get extrinsic incubation period for Ross-Macdonald mosquito model}
\usage{
get_eip_mosquito_RM(model, times)
}
\arguments{
\item{model}{an object from \link{make_MicroMoB}}

\item{times}{vector of times to return}
}
\value{
no return value
}
\description{
Get extrinsic incubation period for Ross-Macdonald mosquito model
}
