% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mosquito_RM_API.R
\name{get_f_mosquito_RM}
\alias{get_f_mosquito_RM}
\title{Get feeding rate for Ross-Macdonald mosquito model}
\usage{
get_f_mosquito_RM(model)
}
\arguments{
\item{model}{an object from \link{make_MicroMoB}}
}
\value{
a vector
}
\description{
Get feeding rate for Ross-Macdonald mosquito model
}
