% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/aquatic_BH_API.R
\name{get_K_aqua_BH}
\alias{get_K_aqua_BH}
\title{Get carrying capacity for Beverton-Holt aquatic mosquito model}
\usage{
get_K_aqua_BH(model, times, places)
}
\arguments{
\item{model}{an object from \link{make_MicroMoB}}

\item{times}{vector of times to get values}

\item{places}{vector of places to get values}
}
\value{
a \link{matrix}
}
\description{
Get carrying capacity for Beverton-Holt aquatic mosquito model
}
