% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/aquatic_trace_API.R
\name{get_lambda_aqua_trace}
\alias{get_lambda_aqua_trace}
\title{Get daily emergence for Beverton-Holt aquatic mosquito model}
\usage{
get_lambda_aqua_trace(model, times, places)
}
\arguments{
\item{model}{an object from \link{make_MicroMoB}}

\item{times}{vector of times to get values}

\item{places}{vector of places to get values}
}
\value{
a \link{matrix}
}
\description{
Get daily emergence for Beverton-Holt aquatic mosquito model
}
