% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mosquito_RM_API.R
\name{get_nu_mosquito_RM}
\alias{get_nu_mosquito_RM}
\title{Get number of eggs laid per oviposition for Ross-Macdonald mosquito model}
\usage{
get_nu_mosquito_RM(model)
}
\arguments{
\item{model}{an object from \link{make_MicroMoB}}
}
\value{
a vector
}
\description{
Get number of eggs laid per oviposition for Ross-Macdonald mosquito model
}
