% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mosquito_RM_API.R
\name{get_p_mosquito_RM}
\alias{get_p_mosquito_RM}
\title{Get daily survival probability for Ross-Macdonald mosquito model}
\usage{
get_p_mosquito_RM(model, times, places)
}
\arguments{
\item{model}{an object from \link{make_MicroMoB}}

\item{times}{vector of times to get values}

\item{places}{vector of places to get values}
}
\value{
a matrix
}
\description{
Get daily survival probability for Ross-Macdonald mosquito model
}
