% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mosquito_RM_API.R
\name{get_psi_mosquito_RM}
\alias{get_psi_mosquito_RM}
\title{Get mosquito dispersal matrix for Ross-Macdonald mosquito model}
\usage{
get_psi_mosquito_RM(model)
}
\arguments{
\item{model}{an object from \link{make_MicroMoB}}
}
\value{
a matrix
}
\description{
Get mosquito dispersal matrix for Ross-Macdonald mosquito model
}
