% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ModelObject.R
\name{get_tmax}
\alias{get_tmax}
\title{Get maximum time of simulation from model object}
\usage{
get_tmax(model)
}
\arguments{
\item{model}{an object from \link{make_MicroMoB}}
}
\description{
Get maximum time of simulation from model object
}
