% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ModelObject.R
\name{get_tnow}
\alias{get_tnow}
\title{Get current time of simulation from model object}
\usage{
get_tnow(model)
}
\arguments{
\item{model}{an object from \link{make_MicroMoB}}
}
\description{
Get current time of simulation from model object
}
