% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils_internal.R
\name{is_binary}
\alias{is_binary}
\title{Does a numeric object consist of only zeros and ones?}
\usage{
is_binary(x)
}
\arguments{
\item{x}{a \link{numeric} object}
}
\value{
a logical value
}
\description{
Does a numeric object consist of only zeros and ones?
}
