% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/adult-interface.R
\name{list_MYZvars}
\alias{list_MYZvars}
\title{Return the variables as a list}
\usage{
list_MYZvars(y, pars, s)
}
\arguments{
\item{y}{the variables}

\item{pars}{a \link{list}}

\item{s}{the vector species index}
}
\value{
a \link{list}
}
\description{
This method dispatches on the type of \code{pars$MYZpar[[s]]}.
}
