% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/human-SIS.R
\name{list_Xvars.SIS}
\alias{list_Xvars.SIS}
\title{Return the variables as a list}
\usage{
\method{list_Xvars}{SIS}(y, pars, i)
}
\arguments{
\item{y}{the variables}

\item{pars}{a \link{list}}

\item{i}{the host species index}
}
\value{
a \link{list}
}
\description{
This method dispatches on the type of \code{pars$Xpar}
}
