% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/adult-ENBRQ.R
\name{make_inits_MYZ_ENBRQ}
\alias{make_inits_MYZ_ENBRQ}
\title{Make inits for ENBRQ adult mosquito model}
\usage{
make_inits_MYZ_ENBRQ(pars, E0, N0, B0, Q0, R10, R20)
}
\arguments{
\item{pars}{a \link{list}}

\item{E0}{total mosquito density at each patch}

\item{N0}{total mosquito density at each patch}

\item{B0}{blood feeding mosquito density at each patch}

\item{Q0}{egg laying mosquito density at each patch}

\item{R10}{total stage 1 resting mosquitoes in each patch}

\item{R20}{total stage 2 resting mosquitoes in each patch}
}
\value{
a \link{list}
}
\description{
Make inits for ENBRQ adult mosquito model
}
