% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/adult-lumpy.R
\name{make_inits_MYZ_RMlumpy}
\alias{make_inits_MYZ_RMlumpy}
\title{Make inits for RMlumpy adult mosquito model}
\usage{
make_inits_MYZ_RMlumpy(pars, M0, P0, U0, Y0, Z0)
}
\arguments{
\item{pars}{a \link{list}}

\item{M0}{total mosquito density at each patch}

\item{P0}{total parous mosquito density at each patch}

\item{U0}{total uninfected mosquito density at each patch}

\item{Y0}{infected mosquito density at each patch}

\item{Z0}{infectious mosquito density at each patch}
}
\value{
a \link{list}
}
\description{
Make inits for RMlumpy adult mosquito model
}
