% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ModelObject.R
\name{make_MicroMoB}
\alias{make_MicroMoB}
\title{Make a model object}
\usage{
make_MicroMoB(tmax, p)
}
\arguments{
\item{tmax}{number of days to simulate}

\item{p}{number of places}
}
\description{
The model object is a hashed \link{environment}. By default it contains
a single list, \code{model$global} storing global state.
}
