% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/adult-ENBRQ.R
\name{make_MYZinits_ENBRQ}
\alias{make_MYZinits_ENBRQ}
\title{Make inits for ENBRQ adult mosquito model}
\usage{
make_MYZinits_ENBRQ(
  nPatches,
  nR1,
  MYZopts = list(),
  E0 = 5,
  N0 = 5,
  B0 = 5,
  Q0 = 1,
  R10 = 0,
  R20 = 0
)
}
\arguments{
\item{nPatches}{the number of patches in the model}

\item{nR1}{number of time steps in R1}

\item{MYZopts}{a \link{list} of values that overwrites the defaults}

\item{E0}{total null-parous, blood feeding mosquito density}

\item{N0}{total null-parous, blood feeding mosquito density}

\item{B0}{total parous, blood feeding mosquito density at each patch}

\item{Q0}{total egg laying mosquito density at each patch}

\item{R10}{resting, stage 1, mosquito density at each patch}

\item{R20}{resting, stage 2, mosquito density at each patch}
}
\value{
a \link{list}
}
\description{
Make inits for ENBRQ adult mosquito model
}
