% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/adult-ENBRQ.R
\name{make_MYZpar_ENBRQ}
\alias{make_MYZpar_ENBRQ}
\title{Make parameters for ENBRQ adult mosquito model}
\usage{
make_MYZpar_ENBRQ(
  nPatches,
  MYZopts = list(),
  calK,
  D = 4,
  nR1 = 3,
  p = 11/12,
  sigma = 1/8,
  f = 0.3,
  q = 0.95,
  r2q = 0.5,
  nu = 1,
  eggsPerBatch = 60,
  p_mod = "dddn",
  sigma_mod = "dddn",
  f_mod = "dddn",
  q_mod = "dddn",
  nu_mod = "dddn"
)
}
\arguments{
\item{nPatches}{is the number of patches, an integer}

\item{MYZopts}{a \link{list} of values that overwrites the defaults}

\item{calK}{a mosquito dispersal matrix of dimensions \code{nPatches} by \code{nPatches}}

\item{D}{number of time steps per day}

\item{nR1}{number of time steps in R1}

\item{p}{daily mosquito survival}

\item{sigma}{emigration rate}

\item{f}{feeding rate}

\item{q}{human blood fraction}

\item{r2q}{proportion of R2 transitioning to Q}

\item{nu}{oviposition rate, per mosquito}

\item{eggsPerBatch}{eggs laid per oviposition}

\item{p_mod}{a name to dispatch F_p}

\item{sigma_mod}{a name to dispatch F_sigma}

\item{f_mod}{a name to dispatch F_f}

\item{q_mod}{a name to dispatch F_q}

\item{nu_mod}{a name to dispatch F_nu}
}
\value{
a \link{list}
}
\description{
Make parameters for ENBRQ adult mosquito model
}
