% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/adult-lumpy.R
\name{make_MYZpar_RMlumpy}
\alias{make_MYZpar_RMlumpy}
\title{Make parameters for RMlumpy ODE adult mosquito model}
\usage{
make_MYZpar_RMlumpy(
  nPatches,
  MYZopts = list(),
  EIPname,
  calK,
  p = 11/12,
  sigma = 1/8,
  f = 0.3,
  q = 0.95,
  eip = 12,
  nu = 1,
  eggsPerBatch = 60,
  p_mod = "static",
  sigma_mod = "static",
  f_mod = "static",
  q_mod = "static",
  nu_mod = "static"
)
}
\arguments{
\item{nPatches}{is the number of patches, an integer}

\item{MYZopts}{a \link{list} of values that overwrites the defaults}

\item{EIPname}{a string: the class name for the EIP model}

\item{calK}{a mosquito dispersal matrix of dimensions \code{nPatches} by \code{nPatches}}

\item{p}{daily mosquito survival}

\item{sigma}{emigration rate}

\item{f}{feeding rate}

\item{q}{human blood fraction}

\item{eip}{the maximum number of cohorts in the EIP}

\item{nu}{oviposition rate, per mosquito}

\item{eggsPerBatch}{eggs laid per oviposition}

\item{p_mod}{a name to dispatch F_p}

\item{sigma_mod}{a name to dispatch F_sigma}

\item{f_mod}{a name to dispatch F_f}

\item{q_mod}{a name to dispatch F_q}

\item{nu_mod}{a name to dispatch F_nu}
}
\value{
a \link{list}
}
\description{
Make parameters for RMlumpy ODE adult mosquito model
}
