% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/adult-interface.R
\name{make_Omega}
\alias{make_Omega}
\title{Make the mosquito demography matrix}
\usage{
make_Omega(p, sigma, K, nPatches)
}
\arguments{
\item{p}{proportion surviving}

\item{sigma}{fraction emigrating}

\item{K}{mosquito dispersal matrix}

\item{nPatches}{number of patches}
}
\value{
a \link{matrix} of dimensions \code{nPatches} by \code{nPatches}
}
\description{
Make the mosquito demography matrix
}
