% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dts_utils.R
\name{make_parameters_dts}
\alias{make_parameters_dts}
\title{Make base parameters, assuming nVectors = nHosts = 1}
\usage{
make_parameters_dts()
}
\value{
a \link{list}
}
\description{
Make base parameters, assuming nVectors = nHosts = 1
}
