% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/adult-basicM.R
\name{make_parameters_MYZ_basicM}
\alias{make_parameters_MYZ_basicM}
\title{Make parameters for basicM adult mosquito model}
\usage{
make_parameters_MYZ_basicM(pars, p, sigma, f, q, nu, eggsPerBatch, calK)
}
\arguments{
\item{pars}{a \link{list}}

\item{p}{daily mosquito survival}

\item{sigma}{emigration rate}

\item{f}{feeding rate}

\item{q}{human blood fraction}

\item{nu}{oviposition rate, per mosquito}

\item{eggsPerBatch}{eggs laid per oviposition}

\item{calK}{mosquito dispersal matrix of dimensions \code{nPatches} by \code{nPatches}}
}
\value{
a \link{list}
}
\description{
Make parameters for basicM adult mosquito model
}
