% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/adult-ENBRQ.R
\name{make_parameters_MYZ_ENBRQ}
\alias{make_parameters_MYZ_ENBRQ}
\title{Make parameters for ENBRQ adult mosquito model}
\usage{
make_parameters_MYZ_ENBRQ(
  pars,
  D,
  nR1,
  p,
  sigma,
  f,
  q,
  r2q,
  nu,
  eggsPerBatch,
  calK
)
}
\arguments{
\item{pars}{a \link{list}}

\item{D}{time steps per day}

\item{nR1}{time steps per day}

\item{p}{daily mosquito survival}

\item{sigma}{emigration rate}

\item{f}{feeding rate}

\item{q}{human blood fraction}

\item{r2q}{proportion of R2 transitioning to Q}

\item{nu}{oviposition rate, per mosquito}

\item{eggsPerBatch}{eggs laid per oviposition}

\item{calK}{mosquito dispersal matrix of dimensions \code{nPatches} by \code{nPatches}}
}
\value{
a \link{list}
}
\description{
Make parameters for ENBRQ adult mosquito model
}
