% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MicroMoB.R
\docType{package}
\name{MicroMoB}
\alias{MicroMoB}
\title{MicroMoB: Discrete time simulation of mosquito-borne pathogen transmission}
\description{
This package does things.
}
\section{MicroMoB functions}{

The MicroMoB functions ...
}

