% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mosquito_RM_API.R
\name{modify_kappa_mosquito_RM}
\alias{modify_kappa_mosquito_RM}
\title{Modify kappa for Ross-Macdonald mosquito model}
\usage{
modify_kappa_mosquito_RM(model, kappa)
}
\arguments{
\item{model}{an object from \link{make_MicroMoB}}

\item{kappa}{new value of \code{kappa}}
}
\value{
no return value
}
\description{
Change \code{kappa}.
}
