% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mosquito_RM_API.R
\name{modify_nu_mosquito_RM}
\alias{modify_nu_mosquito_RM}
\title{Modify number of eggs laid per oviposition for Ross-Macdonald mosquito model}
\usage{
modify_nu_mosquito_RM(model, nu)
}
\arguments{
\item{model}{an object from \link{make_MicroMoB}}

\item{nu}{new number of eggs laid per oviposition}
}
\value{
no return value
}
\description{
Change the number of eggs laid per oviposition parameter \code{nu}.
}
