% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mosquito_RM_API.R
\name{modify_p_mosquito_RM}
\alias{modify_p_mosquito_RM}
\title{Modify daily survival probability for Ross-Macdonald mosquito model}
\usage{
modify_p_mosquito_RM(model, p, times, places)
}
\arguments{
\item{model}{an object from \link{make_MicroMoB}}

\item{p}{new human blood feeding fraction}

\item{times}{vector of times to set the new values}

\item{places}{vector of places to set the new values}
}
\value{
no return value
}
\description{
Change the daily survival probability parameter \code{p} for some times
and places. The parameter \code{p} is stored internally as a matrix so that \code{times}
and \code{places} are used to modify a submatrix, therefore the new value \code{p} should
either be a scalar value to update the entire submatrix or a matrix of \code{places} rows
and \code{times} columns.
}
