% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/observation.R
\name{observe_pfpr}
\alias{observe_pfpr}
\title{Observe PfPR in human strata}
\usage{
observe_pfpr(model, parameters)
}
\arguments{
\item{model}{an object from \link{make_MicroMoB}}

\item{parameters}{a named \link{list}, should have elements \code{sens} (sensitivity),
\code{spec} (specificity), and a vector of length equal to number of strata \code{testprop}
which gives the proportion of each strata to be tested.}
}
\value{
an \link{array} of counts, with actual condition as first dimension and tested condition
as the second dimension, and the third dimension is the human strata
}
\description{
This method dispatches on the type of \code{model$human}.
}
