% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/observation_humans.R
\name{observe_prev.MOI}
\alias{observe_prev.MOI}
\title{Observe prevalence in human strata for MOI model}
\usage{
\method{observe_prev}{MOI}(model, parameters)
}
\arguments{
\item{model}{an object from \link{make_MicroMoB}}

\item{parameters}{a named \link{list}, should have elements \code{sens} (sensitivity),
\code{spec} (specificity), and a vector of length equal to number of strata \code{testprop}
which gives the proportion of each strata to be tested.}
}
\value{
an \link{array} of counts, with actual condition as first dimension and tested condition
as the second dimension, and the third dimension is the human strata
}
\description{
Observe prevalence in human strata for MOI model
}
