% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/aquatic_BH.R
\name{output_aqua.BH}
\alias{output_aqua.BH}
\title{Get output for aquatic (immature) mosquito populations with Beverton-Holt dynamics}
\usage{
\method{output_aqua}{BH}(model)
}
\arguments{
\item{model}{an object from \link{make_MicroMoB}}
}
\value{
a \link{data.frame} with columns \code{L} (immature) and \code{A} (emerging pupae)
}
\description{
Return a \link{data.frame}.
}
