% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/aquatic_interface.R
\name{output_aqua}
\alias{output_aqua}
\title{Get output for aquatic (immature) mosquito populations}
\usage{
output_aqua(model)
}
\arguments{
\item{model}{an object from \link{make_MicroMoB}}
}
\value{
a \link{data.frame}
}
\description{
This method dispatches on the type of \code{model$aqua}. It returns
the current state of the aquatic component.
}
