% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/aquatic_trace.R
\name{output_aqua.trace}
\alias{output_aqua.trace}
\title{Get output for aquatic (immature) mosquito populations with forced emergence}
\usage{
\method{output_aqua}{trace}(model)
}
\arguments{
\item{model}{an object from \link{make_MicroMoB}}
}
\value{
a \link{data.frame}
}
\description{
This function returns an empty \link{data.frame} as trace models do
not have endogenous dynamics.
}
