% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mosquito_interface.R
\name{output_mosquitoes}
\alias{output_mosquitoes}
\title{Get output for mosquito populations}
\usage{
output_mosquitoes(model)
}
\arguments{
\item{model}{an object from \link{make_MicroMoB}}
}
\value{
a \link{data.frame}
}
\description{
This method dispatches on the type of \code{model$mosquito}. It returns
the current state of the adult mosquito component.
}
