% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mosquito_trace.R
\name{output_mosquitoes.trace}
\alias{output_mosquitoes.trace}
\title{Get output for null mosquito populations}
\usage{
\method{output_mosquitoes}{trace}(model)
}
\arguments{
\item{model}{an object from \link{make_MicroMoB}}
}
\value{
a \link{data.frame}
}
\description{
This function returns an empty \link{data.frame} as trace models do
not have endogeneous dynamics.
}
