% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dts_utils.R
\name{parse_dts_out}
\alias{parse_dts_out}
\title{Parse the output of an object returned by deSolve}
\usage{
parse_dts_out(dts_out, pars)
}
\arguments{
\item{dts_out}{a \link{matrix} of orbits returned by deSolve}

\item{pars}{a \link{list}}
}
\value{
varslist a \link{list}
}
\description{
Parse the output of an object returned by deSolve
}
