% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/aquatic-basic.R
\name{parse_dts_out_L.basic}
\alias{parse_dts_out_L.basic}
\title{Parse the variable names for the basic model}
\usage{
\method{parse_dts_out_L}{basic}(dts_out, pars, s)
}
\arguments{
\item{dts_out}{a \link{matrix} of outputs from deSolve}

\item{pars}{a \link{list} that defines the model}

\item{s}{the species index}
}
\value{
\link{list}
}
\description{
Implements \link{parse_dts_out_L} for basic competition model.
}
