% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/adult-ENBRQ.R
\name{parse_dts_out_MYZ.ENBRQ}
\alias{parse_dts_out_MYZ.ENBRQ}
\title{Parse the output of deSolve and return variables for the ENBRQ model}
\usage{
\method{parse_dts_out_MYZ}{ENBRQ}(dts_out, pars, s)
}
\arguments{
\item{dts_out}{a \link{matrix} of outputs from deSolve}

\item{pars}{a \link{list} that defines a model}

\item{s}{the species index}
}
\value{
a \link{list}
}
\description{
Implements \link{parse_dts_out_MYZ} for the ENBRQ model
}
