% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dts_utils.R
\name{parse_dts_out_vec}
\alias{parse_dts_out_vec}
\title{Parse the output of an object returned by deSolve}
\usage{
parse_dts_out_vec(vec, pars)
}
\arguments{
\item{vec}{a \link{vector} with the variables, as returned by rootsolve}

\item{pars}{a \link{list}}
}
\value{
varslist a \link{list}
}
\description{
Parse the output of an object returned by deSolve
}
