% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/api_mosquito.R
\name{run_api_mosquito}
\alias{run_api_mosquito}
\title{Run Plumber API for mosquito-only simulation}
\usage{
run_api_mosquito(...)
}
\arguments{
\item{...}{arguments passed to \link[plumber:pr_run]{plumber::pr_run}}
}
\description{
Run Plumber API for mosquito-only simulation
}
