% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mosquito_RM_API.R
\name{set_eip_mosquito_RM}
\alias{set_eip_mosquito_RM}
\title{Set extrinsic incubation period for Ross-Macdonald mosquito model}
\usage{
set_eip_mosquito_RM(model, eip, times)
}
\arguments{
\item{model}{an object from \link{make_MicroMoB}}

\item{eip}{new extrinsic incubation period values}

\item{times}{vector of times to set the new values}
}
\value{
no return value
}
\description{
Change the extrinsic incubation period parameter \code{eip} for some
set of times. The new values \code{eip} should either be a scalar or a vector
of length equal to the length of \code{times}.
}
