% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mosquito_RM_API.R
\name{set_f_mosquito_RM}
\alias{set_f_mosquito_RM}
\title{Set feeding rate for Ross-Macdonald mosquito model}
\usage{
set_f_mosquito_RM(model, f)
}
\arguments{
\item{model}{an object from \link{make_MicroMoB}}

\item{f}{new blood feeding rate}
}
\value{
no return value
}
\description{
Change the feeding rate parameter \code{f}.
}
