% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/aquatic_trace_API.R
\name{set_lambda_aqua_trace}
\alias{set_lambda_aqua_trace}
\title{Set daily emergence for trace (forced) aquatic mosquito model}
\usage{
set_lambda_aqua_trace(model, lambda, times, places)
}
\arguments{
\item{model}{an object from \link{make_MicroMoB}}

\item{lambda}{new emergence}

\item{times}{vector of times to set the new values}

\item{places}{vector of places to set the new values}
}
\value{
no return value
}
\description{
Change the daily emergence parameter \code{lambda} for some times
and places. The parameter \code{lambda} is stored internally as a matrix so that \code{times}
and \code{places} are used to modify a submatrix, therefore the new value \code{lambda} should
either be a scalar value to update the entire submatrix or a matrix of \code{places} rows
and \code{times} columns.
}
