% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/aquatic_BH_API.R
\name{set_molt_aqua_BH}
\alias{set_molt_aqua_BH}
\title{Set daily maturation probability for Beverton-Holt aquatic mosquito model}
\usage{
set_molt_aqua_BH(model, molt, times, places)
}
\arguments{
\item{model}{an object from \link{make_MicroMoB}}

\item{molt}{new daily maturation probability}

\item{times}{vector of times to set the new values}

\item{places}{vector of places to set the new values}
}
\value{
no return value
}
\description{
Change the daily maturation probability parameter \code{molt} for some times
and places. The parameter \code{molt} is stored internally as a matrix so that \code{times}
and \code{places} are used to modify a submatrix, therefore the new value \code{molt} should
either be a scalar value to update the entire submatrix or a matrix of \code{places} rows
and \code{times} columns.
}
