% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mosquito_RM_API.R
\name{set_psi_mosquito_RM}
\alias{set_psi_mosquito_RM}
\title{Set mosquito dispersal matrix for Ross-Macdonald mosquito model}
\usage{
set_psi_mosquito_RM(model, psi)
}
\arguments{
\item{model}{an object from \link{make_MicroMoB}}

\item{psi}{new mosquito dispersal matrix}
}
\value{
no return value
}
\description{
Change the mosquito dispersal matrix parameter \code{psi}.
}
