% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mosquito_RM_API.R
\name{set_q_mosquito_RM}
\alias{set_q_mosquito_RM}
\title{Set human blood feeding fraction for Ross-Macdonald mosquito model}
\usage{
set_q_mosquito_RM(model, q)
}
\arguments{
\item{model}{an object from \link{make_MicroMoB}}

\item{q}{new human blood feeding fraction}
}
\value{
no return value
}
\description{
Change the human blood feeding fraction parameter \code{q}.
}
