% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/aquatic_BH_API.R
\name{set_surv_aqua_BH}
\alias{set_surv_aqua_BH}
\title{Set daily survival probability for Beverton-Holt aquatic mosquito model}
\usage{
set_surv_aqua_BH(model, surv, times, places)
}
\arguments{
\item{model}{an object from \link{make_MicroMoB}}

\item{surv}{new daily survival probability}

\item{times}{vector of times to set the new values}

\item{places}{vector of places to set the new values}
}
\value{
no return value
}
\description{
Change the daily survival probability parameter \code{surv} for some times
and places. The parameter \code{surv} is stored internally as a matrix so that \code{times}
and \code{places} are used to modify a submatrix, therefore the new value \code{surv} should
either be a scalar value to update the entire submatrix or a matrix of \code{places} rows
and \code{times} columns.
}
