% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Human.R
\name{setup.human.strata}
\alias{setup.human.strata}
\title{Setup a human model with strata}
\usage{
\method{setup.human}{strata}(type, model, H, J = NULL, ...)
}
\arguments{
\item{type}{a character in \code{c("simple", "strata")}}

\item{model}{a model object (an \link{environment})}

\item{H}{a vector of human population sizes}

\item{J}{a matrix whose columns assign human strata to patches (rows); the
columns must all sum to one. If \code{J} is \code{NULL} then a diagonal matrix of ones
is used; this assumes that the only level of population stratification is
patch of residence.}

\item{...}{other arguments to be passed to type methods}
}
\description{
This sets up a human model object.
}
