% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/other_hosts.R
\name{setup.otherhosts}
\alias{setup.otherhosts}
\title{Setup other blood hosts}
\usage{
setup.otherhosts(type, model, ...)
}
\arguments{
\item{type}{a character in \code{c("simple")}}

\item{model}{a model object (an \link{environment})}

\item{...}{other arguments to be passed to type methods}
}
\description{
Setup a other blood hosts (vertebrate mammals) object.
The \code{model} object will have a list named \code{otherhosts} added to it.
}
\section{strata}{
 see \link{setup.human.strata}
}

