% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/other_trace.R
\name{setup_alternative_trace}
\alias{setup_alternative_trace}
\title{Setup trace driven alternative blood hosts}
\usage{
setup_alternative_trace(model, O = NULL)
}
\arguments{
\item{model}{an object from \link{make_MicroMoB}}

\item{O}{a time varying trace passed to \link{time_patch_varying_parameter}
or \code{NULL} to set to \code{0} (no alternative blood hosts)}
}
\value{
no return value
}
\description{
This model complies with the visitors component interface. It adds
a named list \code{model$alternative}.
}
