% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/aquatic_BH.R
\name{setup_aqua_BH}
\alias{setup_aqua_BH}
\title{Setup aquatic (immature) mosquito model with Beverton-Holt dynamics}
\usage{
setup_aqua_BH(model, molt, surv, K, stochastic)
}
\arguments{
\item{model}{an object from \link{make_MicroMoB}}

\item{molt}{proportion of immature stages which will mature and emerge as adults each day}

\item{surv}{daily survival probability}

\item{K}{carrying capacity}

\item{stochastic}{should the model update deterministically or stochastically?}
}
\description{
A single compartment for all aquatic stages is modeled which
suffers density dependent mortality like the Beverton-Holt model.
}
\details{
All parameters can be passed either as a vector of length equal to \code{p}, a matrix with \code{p} rows
and \code{tmax} columns, or a matrix with \code{p} rows and \code{365} columns.
}
