% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/aquatic_trace.R
\name{setup_aqua_trace}
\alias{setup_aqua_trace}
\title{Setup aquatic (immature) mosquito model with trace (forced) emergence}
\usage{
setup_aqua_trace(model, lambda, stochastic)
}
\arguments{
\item{model}{an object from \link{make_MicroMoB}}

\item{lambda}{either a vector of length equal to \code{p}, a matrix with \code{p} rows
and \code{tmax} columns, or a matrix with \code{p} rows and \code{365} columns}

\item{stochastic}{should the model update deterministically or stochastically?}
}
\description{
Emergence is passed as a (possibly time varying) parameter which is
decoupled from the adult mosquito dynamics.
}
