% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/human.R
\name{setup_biteweight.simple}
\alias{setup_biteweight.simple}
\title{Setup simple biting weight model}
\usage{
\method{setup_biteweight}{simple}(type, model, wt = NULL, ...)
}
\arguments{
\item{type}{a character in \code{c("simple")}}

\item{model}{a model object (an \link{environment})}

\item{wt}{optional vector of biting weights, if \code{NULL} use 1 for all strata}

\item{...}{other arguments to be passed to type methods}
}
\description{
A simple vector of biting weights that do not change over time.
}
