% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/eip_fixed.R
\name{setup_EIP.fixed}
\alias{setup_EIP.fixed}
\title{Set up a fixed model for the EIP}
\usage{
\method{setup_EIP}{fixed}(EIPname, MYZpar, MYZopts = list())
}
\arguments{
\item{EIPname}{the class name of the function}

\item{MYZpar}{the MYZ parameters}

\item{MYZopts}{is a \link{list} that overwrites default options}
}
\value{
\link{list}
}
\description{
Set up a fixed model for the EIP
}
