% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/eip.R
\name{setup_EIP}
\alias{setup_EIP}
\title{Set up the fixed model for control forcing (do nothing)}
\usage{
setup_EIP(EIPname, MYZpar, MYZopts = list())
}
\arguments{
\item{EIPname}{the class name of the function}

\item{MYZpar}{the MYZ parameters}

\item{MYZopts}{is a \link{list} that overwrites default options}
}
\value{
\link{list} MYZpar with the EIPmod attached
}
\description{
Set up the fixed model for control forcing (do nothing)
}
