% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/eip_fixed.R
\name{setup_eip_fixed}
\alias{setup_eip_fixed}
\title{Set up a fixed model for the EIP}
\usage{
setup_eip_fixed(MYZopts = list(), MYZpar)
}
\arguments{
\item{MYZopts}{a \link{list}}

\item{MYZpar}{the MYZ parameters}
}
\value{
\link{list}
}
\description{
Set up a fixed model for the EIP
}
