% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/eip_null.R
\name{setup_eip_null}
\alias{setup_eip_null}
\title{Set up a null model for the EIP}
\usage{
setup_eip_null(MYZopts = list(), MYZpar)
}
\arguments{
\item{MYZopts}{a \link{list}}

\item{MYZpar}{the MYZ parameters}
}
\value{
\link{list}
}
\description{
Set up a null model for the EIP
}
