% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/humans_SIP.R
\name{setup_humans_SIP}
\alias{setup_humans_SIP}
\title{Setup humans with SIP pathogen model}
\usage{
setup_humans_SIP(
  model,
  stochastic,
  theta,
  wf = NULL,
  SIP,
  b = 0.55,
  c = 0.15,
  r = 1/200,
  rho = 0.07,
  eta = 1/32
)
}
\arguments{
\item{model}{an object from \link{make_MicroMoB}}

\item{stochastic}{should the model update deterministically or stochastically?}

\item{theta}{a time spent matrix}

\item{wf}{biting weights}

\item{SIP}{matrix of strata (rows) by health states (SIP)}

\item{b}{transmission efficiency (mosquito to human)}

\item{c}{transmission efficiency (human to mosquito)}

\item{r}{recovery rate (inverse of infectious duration)}

\item{rho}{probability of treatment upon infection}

\item{eta}{rate at which prophylaxis decays}
}
\value{
no return value
}
\description{
A simple SIP (Susceptible-Infected-Protected) model
}
