% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/humans_SIR.R
\name{setup_humans_SIR}
\alias{setup_humans_SIR}
\title{Setup humans with SIR infection model}
\usage{
setup_humans_SIR(
  model,
  stochastic,
  theta,
  wf = NULL,
  H,
  SIR,
  b = 0.55,
  c = 0.15,
  gamma = 1/5
)
}
\arguments{
\item{model}{an object from \link{make_MicroMoB}}

\item{stochastic}{should the model update deterministically or stochastically?}

\item{theta}{a time spent matrix}

\item{wf}{biting weights}

\item{H}{vector of strata population sizes}

\item{SIR}{a matrix giving S, I, R counts (columns) for each strata (rows)}

\item{b}{transmission efficiency (mosquito to human)}

\item{c}{transmission efficiency (human to mosquito)}

\item{gamma}{rate of recovery}
}
\description{
A simple SIR (Susceptible-Infected-Recovered) model
}
