% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/humans_SIS.R
\name{setup_humans_SIS}
\alias{setup_humans_SIS}
\title{Setup humans with SIS pathogen model}
\usage{
setup_humans_SIS(
  model,
  stochastic,
  theta,
  wf = NULL,
  H,
  X,
  b = 0.55,
  c = 0.15,
  r = 1/200
)
}
\arguments{
\item{model}{an object from \link{make_MicroMoB}}

\item{stochastic}{should the model update deterministically or stochastically?}

\item{theta}{a time spent matrix}

\item{wf}{biting weights}

\item{H}{vector of strata population sizes}

\item{X}{number of infectious persons in each strata}

\item{b}{transmission efficiency (mosquito to human)}

\item{c}{transmission efficiency (human to mosquito)}

\item{r}{recovery rate (inverse of infectious duration)}
}
\value{
no return value
}
\description{
A simple SIS (Susceptible-Infected-Susceptible) model
}
