% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mosquito.R
\name{setup_mosquito}
\alias{setup_mosquito}
\title{Setup a mosquito object}
\usage{
setup_mosquito(type, model, ...)
}
\arguments{
\item{type}{a character in \verb{c("rm", "null)}}

\item{model}{a model object (an \link{environment})}

\item{...}{other arguments to be passed to type methods}
}
\description{
Setup a mosquito object. The \code{model} object will have a list named \code{mosquito}
added to it.
}
\seealso{
\link{setup_mosquito.trace}
}
