% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mosquito.R
\name{setup_mosquito.rm_day}
\alias{setup_mosquito.rm_day}
\title{Setup a Ross-Macdonald mosquito object}
\usage{
\method{setup_mosquito}{rm_day}(type, model, mosquito_rm_pars, ...)
}
\arguments{
\item{type}{a character in \code{c("rm_day")}}

\item{model}{a model object (an \link{environment})}

\item{mosquito_rm_pars}{parameters for Ross-Macdonald mosquito model}

\item{...}{other arguments to be passed to type methods}
}
\description{
a model
}
