% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mosquito.R
\name{setup_mosquito.trace}
\alias{setup_mosquito.trace}
\title{Setup a trace mosquito object}
\usage{
\method{setup_mosquito}{trace}(type, model, Z, xi, ...)
}
\arguments{
\item{type}{a character in \verb{c("rm", "null)}}

\item{model}{a model object (an \link{environment})}

\item{Z}{a matrix whose columns are days and rows are patches}

\item{xi}{a}

\item{...}{other arguments to be passed to type methods}
}
\description{
This model simply returns a predefined vector \code{Z}, giving the number
of infectious female mosquitoes ine each patch during each day.
}
