% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mosquito_RM.R
\name{setup_mosquito_RM}
\alias{setup_mosquito_RM}
\title{Setup generalized Ross-Macdonald mosquito model}
\usage{
setup_mosquito_RM(model, stochastic, a, eip, p, psi, M, Y, Z)
}
\arguments{
\item{model}{an object from \link{make_MicroMoB}}

\item{stochastic}{should the model update deterministically or stochastically?}

\item{a}{the feeding rate on humans and birds (normally calculated as \eqn{a = fq} using
Ross-Macdonald parameters)}

\item{eip}{the Extrinsic Incubation Period, may either be a scalar, a vector of
length 365, or a vector of length equal to \code{tmax} in the \code{model} object from \link{make_MicroMoB}}

\item{p}{daily survival probability, may either be a scalar, a vector of
length 365, or a vector of length equal to \code{tmax} in the \code{model} object from \link{make_MicroMoB}}

\item{psi}{a mosquito dispersal matrix (rows must sum to 1)}

\item{M}{total mosquito density per patch}

\item{Y}{density of incubating mosquitoes per patch}

\item{Z}{density of infectious mosquitoes per patch}
}
\description{
This is a generalized RM model which allows for time varying EIP and
survival probability. It complies with the mosquito component interface, and may
be simulated deterministically or stochastically.
}
