% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mosquito_trace.R
\name{setup_mosquito_trace}
\alias{setup_mosquito_trace}
\title{Setup null mosquito model}
\usage{
setup_mosquito_trace(model, oviposit)
}
\arguments{
\item{model}{an object from \link{make_MicroMoB}}

\item{oviposit}{a vector of length \code{p} used as a return value for \link{compute_oviposit}}
}
\description{
This is a null model of mosquito dynamics that is only for testing/verifying aquatic models.
It implements a single method \link{compute_oviposit.trace} and all other methods throw an error.
}
