% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/adult-ENBRQ.R
\name{setup_MYZinits.ENBRQ}
\alias{setup_MYZinits.ENBRQ}
\title{Setup initial values for the ENBRQ model}
\usage{
\method{setup_MYZinits}{ENBRQ}(pars, s, MYZopts = list())
}
\arguments{
\item{pars}{a \link{list}}

\item{s}{the species index}

\item{MYZopts}{a \link{list}}
}
\value{
a \link{list}
}
\description{
Implements \link{setup_MYZinits} for the ENBRQ model
}
