% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/adult-lumpy.R
\name{setup_MYZinits.lumpy}
\alias{setup_MYZinits.lumpy}
\title{Setup initial values for the lumpy model}
\usage{
\method{setup_MYZinits}{lumpy}(pars, s, MYZopts = list())
}
\arguments{
\item{pars}{a \link{list}}

\item{s}{the species index}

\item{MYZopts}{a \link{list}}
}
\value{
a \link{list}
}
\description{
Implements \link{setup_MYZinits} for the lumpy model
}
