% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/adult-Ztrace.R
\name{setup_MYZpar.Ztrace}
\alias{setup_MYZpar.Ztrace}
\title{Setup the Ztrace model}
\usage{
\method{setup_MYZpar}{Ztrace}(MYZname, pars, s, MYZopts = NULL, EIPname = NULL, calK = NULL)
}
\arguments{
\item{MYZname}{the name of the model}

\item{pars}{a \link{list}}

\item{s}{the species index}

\item{MYZopts}{a \link{list}}

\item{EIPname}{is the name of the EIPmod}

\item{calK}{is a \link{matrix}}
}
\value{
a \link{list} vector
}
\description{
Implements \link{setup_MYZpar} for the Ztrace model
}
