% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pathogen.R
\name{setup_pathogen}
\alias{setup_pathogen}
\title{Setup a pathogen model}
\usage{
setup_pathogen(type, model, ...)
}
\arguments{
\item{type}{a character in \code{c("rm")}}

\item{model}{a model object (from \link{setup_model_object})}

\item{...}{other arguments to be passed to type methods}
}
\description{
Setup a time spent model. The model object must have already
been initialized with a human object (see \link{setup_human}). The \code{model} object will have a list named \code{pathogen}
added to it.
}
\seealso{
\link{setup_pathogen.rm}
}
