% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pathogen.R
\name{setup_pathogen.rm}
\alias{setup_pathogen.rm}
\title{Setup daily time spent model}
\usage{
\method{setup_pathogen}{rm}(type, model, rm_parameters, ...)
}
\arguments{
\item{type}{a character in \code{c("rm")}}

\item{model}{a model object (an \link{environment})}

\item{rm_parameters}{a list of parameters}

\item{...}{other arguments to be passed to type methods}
}
\description{
A model of where people spend their time over the day.
}
