% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/time_spent.R
\name{setup_timespent.day}
\alias{setup_timespent.day}
\title{Setup daily time spent model}
\usage{
\method{setup_timespent}{day}(type, model, theta = NULL, ...)
}
\arguments{
\item{type}{a character in \code{c("day")}}

\item{model}{a model object (from \link{setup_model_object})}

\item{theta}{a time spent matrix, if \code{NULL} the identity matrix is used
(everyone stays at their home patch), otherwise \code{theta} should be a
matrix whose rows sum to \eqn{\le 1} giving the daily time spent distribution.}

\item{...}{other arguments to be passed to type methods}
}
\description{
A model of where people spend their time over the day.
}
