% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/time_spent.R
\name{setup_timespent.dt}
\alias{setup_timespent.dt}
\title{Setup fractional day time spent model}
\usage{
\method{setup_timespent}{dt}(type, model, theta, ...)
}
\arguments{
\item{type}{a character in \code{c("day")}}

\item{model}{a model object (an \link{environment})}

\item{theta}{a time spent matrix, if \code{NULL} the identity matrix is used
(everyone stays at their home patch)}

\item{...}{other arguments to be passed to type methods}
}
\description{
Setup fractional day time spent model
}
