% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/time_spent.R
\name{setup_timespent}
\alias{setup_timespent}
\title{Setup a time spent model}
\usage{
setup_timespent(type, model, ...)
}
\arguments{
\item{type}{a character in \code{c("day")}}

\item{model}{a model object (from \link{setup_model_object})}

\item{...}{other arguments to be passed to type methods}
}
\description{
Setup a time spent model. The model object must have already
been initialized with a human object (see \link{setup_human}). This adds
a \link{list} to \code{model$human} named \code{pathogen} (Time spent).
}
\seealso{
\link{setup_timespent.day} \link{setup_timespent.dt}
}
