% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/other_hosts.R
\name{setup_visitors}
\alias{setup_visitors}
\title{Setup visitors}
\usage{
setup_visitors(type, model, ...)
}
\arguments{
\item{type}{a character in \code{c("simple")}}

\item{model}{a model object (from \link{setup_model_object})}

\item{...}{other arguments to be passed to type methods}
}
\description{
Setup a human visitors (vertebrate mammals) object.
These are persons who are present in the spatial domain but are not
residents of any patch (transient visitors).
The \code{model} object will have a list named \code{visitors} added to it.
}
\seealso{
\link{setup_visitors.simple}
}
