% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/other_hosts.R
\name{setup_visitors.simple}
\alias{setup_visitors.simple}
\title{Setup simple visitors}
\usage{
\method{setup_visitors}{simple}(type, model, W_delta = NULL, x_delta = NULL, ...)
}
\arguments{
\item{type}{a character in \code{c("simple")}}

\item{model}{a model object (an \link{environment})}

\item{W_delta}{a vector of blood feeding attempt weights by patch}

\item{x_delta}{a vector of prevalence in visitors by patch}

\item{...}{other arguments to be passed to type methods}
}
\description{
If both parameters are \code{NULL} then
}
