% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/aquatic_BH.R
\name{step_aqua.BH}
\alias{step_aqua.BH}
\title{Update aquatic (immature) mosquito populations for Beverton-Holt dynamics}
\usage{
\method{step_aqua}{BH}(model)
}
\arguments{
\item{model}{an object from \link{make_MicroMoB}}
}
\value{
no return value
}
\description{
This function dispatches on the second class attribute of \code{model$aqua}
for stochastic or deterministic behavior.
}
