% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/aquatic_BH.R
\name{step_aqua.BH_deterministic}
\alias{step_aqua.BH_deterministic}
\title{Update aquatic (immature) mosquito populations for deterministic Beverton-Holt dynamics}
\usage{
\method{step_aqua}{BH_deterministic}(model)
}
\arguments{
\item{model}{an object from \link{make_MicroMoB}}
}
\description{
Run a deterministic state update.
}
