% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/aquatic_BH.R
\name{step_aqua.BH_stochastic}
\alias{step_aqua.BH_stochastic}
\title{Update aquatic (immature) mosquito populations for stochastic Beverton-Holt dynamics}
\usage{
\method{step_aqua}{BH_stochastic}(model)
}
\arguments{
\item{model}{an object from \link{make_MicroMoB}}
}
\description{
Run a stochastic state update.
}
