% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/aquatic_trace.R
\name{step_aqua.trace}
\alias{step_aqua.trace}
\title{Update aquatic (immature) mosquito populations for forced emergence}
\usage{
\method{step_aqua}{trace}(model)
}
\arguments{
\item{model}{an object from \link{make_MicroMoB}}
}
\value{
no return value
}
\description{
This function does nothing as trace models are do not have
endogenous dynamics.
}
