% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/humans_MOI.R
\name{step_humans.MOI_deterministic}
\alias{step_humans.MOI_deterministic}
\title{Update MOI human model (deterministic)}
\usage{
\method{step_humans}{MOI_deterministic}(model)
}
\arguments{
\item{model}{an object from \link{make_MicroMoB}}
}
\value{
no return value
}
\description{
Update MOI human model (deterministic)
}
