% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/humans_MOI.R
\name{step_humans.MOI_stochastic}
\alias{step_humans.MOI_stochastic}
\title{Update MOI human model (stochastic)}
\usage{
\method{step_humans}{MOI_stochastic}(model)
}
\arguments{
\item{model}{an object from \link{make_MicroMoB}}
}
\description{
Update MOI human model (stochastic)
}
