% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/humans_interface.R
\name{step_humans}
\alias{step_humans}
\title{Update human population}
\usage{
step_humans(model)
}
\arguments{
\item{model}{an object from \link{make_MicroMoB}}
}
\value{
no return value
}
\description{
This method dispatches on the type of \code{model$human}.
}
