% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/humans_SIP.R
\name{step_humans.SIP}
\alias{step_humans.SIP}
\title{Update SIP human model}
\usage{
\method{step_humans}{SIP}(model)
}
\arguments{
\item{model}{an object from \link{make_MicroMoB}}
}
\value{
no return value
}
\description{
Update SIP human model
}
