% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/humans_SIP.R
\name{step_humans.SIP_deterministic}
\alias{step_humans.SIP_deterministic}
\title{Update SIP human model (deterministic)}
\usage{
\method{step_humans}{SIP_deterministic}(model)
}
\arguments{
\item{model}{an object from \link{make_MicroMoB}}
}
\value{
no return value
}
\description{
Update SIP human model (deterministic)
}
