% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/humans_SIP.R
\name{step_humans.SIP_stochastic}
\alias{step_humans.SIP_stochastic}
\title{Update SIP human model (stochastic)}
\usage{
\method{step_humans}{SIP_stochastic}(model)
}
\arguments{
\item{model}{an object from \link{make_MicroMoB}}
}
\value{
no return value
}
\description{
Update SIP human model (stochastic)
}
