% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/humans_SIR.R
\name{step_humans.SIR}
\alias{step_humans.SIR}
\title{Update SIR human model}
\usage{
\method{step_humans}{SIR}(model)
}
\arguments{
\item{model}{an object from \link{make_MicroMoB}}
}
\value{
no return value
}
\description{
Update SIR human model
}
