% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/humans_SIR.R
\name{step_humans.SIR_deterministic}
\alias{step_humans.SIR_deterministic}
\title{Update SIR human model (deterministic)}
\usage{
\method{step_humans}{SIR_deterministic}(model)
}
\arguments{
\item{model}{an object from \link{make_MicroMoB}}
}
\description{
Update SIR human model (deterministic)
}
