% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/humans_SIR.R
\name{step_humans.SIR_stochastic}
\alias{step_humans.SIR_stochastic}
\title{Update SIR human model (stochastic)}
\usage{
\method{step_humans}{SIR_stochastic}(model)
}
\arguments{
\item{model}{an object from \link{make_MicroMoB}}
}
\value{
no return value
}
\description{
Update SIR human model (stochastic)
}
