% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/humans_SIS.R
\name{step_humans.SIS_deterministic}
\alias{step_humans.SIS_deterministic}
\title{Update SIS human model (deterministic)}
\usage{
\method{step_humans}{SIS_deterministic}(model)
}
\arguments{
\item{model}{an object from \link{make_MicroMoB}}
}
\description{
Update SIS human model (deterministic)
}
