% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/humans_SIS.R
\name{step_humans.SIS_stochastic}
\alias{step_humans.SIS_stochastic}
\title{Update SIS human model (stochastic)}
\usage{
\method{step_humans}{SIS_stochastic}(model)
}
\arguments{
\item{model}{an object from \link{make_MicroMoB}}
}
\value{
no return value
}
\description{
Update SIS human model (stochastic)
}
