% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mosquito_BQ.R
\name{step_mosquitoes.BQ_deterministic}
\alias{step_mosquitoes.BQ_deterministic}
\title{Update blood feeding & oviposition (BQ) behavioral state mosquitoes (deterministic)}
\usage{
\method{step_mosquitoes}{BQ_deterministic}(model)
}
\arguments{
\item{model}{an object from \link{make_MicroMoB}}
}
\value{
no return value
}
\description{
Update blood feeding & oviposition (BQ) behavioral state mosquitoes (deterministic)
}
