% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mosquito_BQ.R
\name{step_mosquitoes.BQ_stochastic}
\alias{step_mosquitoes.BQ_stochastic}
\title{Update blood feeding & oviposition (BQ) behavioral state mosquitoes (stochastic)}
\usage{
\method{step_mosquitoes}{BQ_stochastic}(model)
}
\arguments{
\item{model}{an object from \link{make_MicroMoB}}
}
\value{
no return value
}
\description{
Update blood feeding & oviposition (BQ) behavioral state mosquitoes (stochastic)
}
