% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mosquito_interface.R
\name{step_mosquitoes}
\alias{step_mosquitoes}
\title{Update mosquito population}
\usage{
step_mosquitoes(model)
}
\arguments{
\item{model}{an object from \link{make_MicroMoB}}
}
\value{
no return value
}
\description{
This method dispatches on the type of \code{model$mosquito}
}
