% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mosquito_RM.R
\name{step_mosquitoes.RM}
\alias{step_mosquitoes.RM}
\title{Update Ross-Macdonald mosquitoes}
\usage{
\method{step_mosquitoes}{RM}(model)
}
\arguments{
\item{model}{an object from \link{make_MicroMoB}}
}
\value{
no return value
}
\description{
This function dispatches on the second argument of \code{model$mosquito}
for stochastic or deterministic behavior.
}
\details{
see \link{step_mosquitoes.RM_deterministic} and \link{step_mosquitoes.RM_stochastic}
}
