% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mosquito_RM.R
\name{step_mosquitoes.RM_deterministic}
\alias{step_mosquitoes.RM_deterministic}
\title{Update Ross-Macdonald mosquitoes (deterministic)}
\usage{
\method{step_mosquitoes}{RM_deterministic}(model)
}
\arguments{
\item{model}{an object from \link{make_MicroMoB}}
}
\value{
no return value
}
\description{
Update Ross-Macdonald mosquitoes (deterministic)
}
