% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mosquito_RM.R
\name{step_mosquitoes.RM_stochastic}
\alias{step_mosquitoes.RM_stochastic}
\title{Update Ross-Macdonald mosquitoes (stochastic)}
\usage{
\method{step_mosquitoes}{RM_stochastic}(model)
}
\arguments{
\item{model}{an object from \link{make_MicroMoB}}
}
\value{
no return value
}
\description{
Update Ross-Macdonald mosquitoes (stochastic)
}
