% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mosquito_trace.R
\name{step_mosquitoes.trace}
\alias{step_mosquitoes.trace}
\title{Update null mosquito population}
\usage{
\method{step_mosquitoes}{trace}(model)
}
\arguments{
\item{model}{an object from \link{make_MicroMoB}}
}
\description{
Update null mosquito population
}
