% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pathogen.R
\name{step_pathogen}
\alias{step_pathogen}
\title{Update the pathogen model}
\usage{
step_pathogen(model, t)
}
\arguments{
\item{model}{a model object (an object from \link{setup_model_object})}

\item{t}{time}
}
\description{
Update the pathogen model
}
\seealso{
\link{step_pathogen.rm}
}
