% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pathogen.R
\name{step_pathogen.rm}
\alias{step_pathogen.rm}
\title{Update the Ross-Macdonald pathogen model}
\usage{
\method{step_pathogen}{rm}(model, t)
}
\arguments{
\item{model}{a model object (an object from \link{setup_model_object})}

\item{t}{time}
}
\description{
Update the Ross-Macdonald pathogen model
}
