% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils_internal.R
\name{time_patch_varying_parameter}
\alias{time_patch_varying_parameter}
\title{Input parameters that may vary by time and patch}
\usage{
time_patch_varying_parameter(param, p, tmax)
}
\arguments{
\item{param}{if given a matrix, it must have nrows equal to \code{p} and ncols equal to
either \code{tmax} or \code{365}; if given a vector it must be of length \code{p}, \code{tmax}, or \code{365}.}

\item{p}{number of patches}

\item{tmax}{number of time steps}
}
\description{
Input parameters that may vary by time and patch
}
