% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils_internal.R
\name{time_varying_parameter}
\alias{time_varying_parameter}
\title{Input parameters that may vary by time}
\usage{
time_varying_parameter(param, tmax)
}
\arguments{
\item{param}{a vector of length \code{1}, \code{tmax}, or \code{365}.}

\item{tmax}{number of time steps}
}
\description{
Input parameters that may vary by time
}
