% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/adult-basicM.R
\name{update_inits_MYZ.basicM}
\alias{update_inits_MYZ.basicM}
\title{Make inits for basicM adult mosquito model}
\usage{
\method{update_inits_MYZ}{basicM}(pars, y0, s)
}
\arguments{
\item{pars}{a \link{list}}

\item{y0}{a vector of variable values from a simulation}

\item{s}{the species index}
}
\value{
a \link{list}
}
\description{
Make inits for basicM adult mosquito model
}
