% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/adult-lumpy.R
\name{update_inits_MYZ.RMlumpy}
\alias{update_inits_MYZ.RMlumpy}
\title{Make inits for RMlumpy adult mosquito model}
\usage{
\method{update_inits_MYZ}{RMlumpy}(pars, y0, s)
}
\arguments{
\item{pars}{a \link{list}}

\item{y0}{a vector of variable values from a simulation}

\item{s}{the species index}
}
\value{
a \link{list}
}
\description{
Make inits for RMlumpy adult mosquito model
}
