% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/blood_feeding.R
\name{compute_beta.dt}
\alias{compute_beta.dt}
\title{Compute fractional biting distribution array (\eqn{\beta})}
\usage{
\method{compute_beta}{dt}(human, xi, t)
}
\arguments{
\item{human}{an object from \link{setup_human}}

\item{xi}{probabilities to initiate a feeding search during each fraction
of the day (must sum to 1)}

\item{t}{time}
}
\value{
an array of dimension \eqn{n \times p \times d}
}
\description{
Compute fractional biting distribution array (\eqn{\beta})
}
