% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/blood_feeding.R
\name{compute_beta}
\alias{compute_beta}
\title{Compute biting distribution matrix (\eqn{\beta})}
\usage{
compute_beta(human, xi, t)
}
\arguments{
\item{human}{an object from \link{setup_human}}

\item{xi}{probabilities to initiate a feeding search during each fraction
of the day (must sum to 1)}

\item{t}{time}
}
\description{
Compute biting distribution matrix (\eqn{\beta})
}
\seealso{
\link{compute_beta.day} \link{compute_beta.dt}
}
