% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bloodmeal.R
\name{compute_bloodmeal}
\alias{compute_bloodmeal}
\title{Compute bloodmeals taken by mosquitoes on hosts}
\usage{
compute_bloodmeal(model)
}
\arguments{
\item{model}{an object from \link{make_MicroMoB}}
}
\description{
This should be run prior to any \code{step} functions to update
components over a time step. It computes various quantities related to
disease transmission between species using the generic interfaces (methods)
provided by each component. It updates \code{h}, vectors giving per-capita force
of infection for the bird and human components, and \code{kappa}, the net infectiousness
of hosts for the mosquito component.
}
