% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/aquatic_BH.R
\name{compute_emergents.BH}
\alias{compute_emergents.BH}
\title{Compute number of newly emerging adults from Beverton-Holt dynamics}
\usage{
\method{compute_emergents}{BH}(model)
}
\arguments{
\item{model}{an object from \link{make_MicroMoB}}
}
\value{
a vector of length \code{p} giving the number of newly emerging adult in each patch
}
\description{
This function dispatches on the second class attribute of \code{model$aqua}
for stochastic or deterministic behavior.
}
