% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mosquito_BQ.R
\name{compute_f.BQ}
\alias{compute_f.BQ}
\title{Compute mosquito feeding rate for BQ model (\eqn{f})}
\usage{
\method{compute_f}{BQ}(model, B)
}
\arguments{
\item{model}{an object from \link{make_MicroMoB}}

\item{B}{a vector of length \code{p} giving total blood host availability by patch}
}
\value{
a vector of length \code{p} giving the per-capita blood feeding rate of mosquitoes in each blood feeding haunt
}
\description{
Compute mosquito feeding rate for BQ model (\eqn{f})
}
