% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/humans_interface.R
\name{compute_H}
\alias{compute_H}
\title{Compute human population strata sizes (\eqn{H})}
\usage{
compute_H(model)
}
\arguments{
\item{model}{an object from \link{make_MicroMoB}}
}
\value{
a vector of length \code{n} giving the size of each human population stratum
}
\description{
This method dispatches on the type of \code{model$human}.
}
