% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mosquito_interface.R
\name{compute_oviposit}
\alias{compute_oviposit}
\title{Compute number of eggs laid from oviposition for each patch}
\usage{
compute_oviposit(model)
}
\arguments{
\item{model}{an object from \link{make_MicroMoB}}
}
\description{
This method dispatches on the type of \code{model$mosquito}
}
