% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mosquito_RM.R
\name{compute_oviposit.RM}
\alias{compute_oviposit.RM}
\title{Compute number of eggs laid from oviposition for each patch for RM model}
\usage{
\method{compute_oviposit}{RM}(model)
}
\arguments{
\item{model}{an object from \link{make_MicroMoB}}
}
\value{
a vector of length \code{p} giving the total number of eggs laid by adult mosquitoes in each patch
}
\description{
This method returns a vector of length \code{p}.
}
\details{
see \link{compute_oviposit.RM_deterministic} and \link{compute_oviposit.RM_stochastic}
}
