% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/aquatic_trace.R
\name{compute_oviposit.trace}
\alias{compute_oviposit.trace}
\title{Compute number of eggs laid from oviposition for each patch}
\usage{
\method{compute_oviposit}{trace}(model)
}
\arguments{
\item{model}{an object from \link{make_MicroMoB}}
}
\description{
This is to be used with modeling emergence as a trace.
}
