% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/time_spent.R
\name{compute_Psi}
\alias{compute_Psi}
\title{Compute time at risk (\eqn{\Psi^{T}})}
\usage{
compute_Psi(human, xi, t)
}
\arguments{
\item{human}{an object from \link{setup_human}}

\item{xi}{probabilities to initiate a feeding search during each fraction
of the day (must sum to 1)}

\item{t}{time (day)}
}
\value{
a (transposed) time at risk \link{matrix} or \link{array}, which should be attached to the \code{timespent}
object and stored for later computation (human availability and the biting distribution matrix)
}
\description{
Dispatch is done on the class of \code{human$timespent}
}
\seealso{
\link{compute_Psi.day} \link{compute_Psi.dt}
}
