% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/humans_SIR.R
\name{compute_Psi.SIR}
\alias{compute_Psi.SIR}
\title{Compute time at risk matrix for SIR model (\eqn{\Psi})}
\usage{
\method{compute_Psi}{SIR}(model)
}
\arguments{
\item{model}{an object from \link{make_MicroMoB}}
}
\value{
a matrix with \code{n} rows and \code{p} columns, the time at risk matrix
}
\description{
Compute time at risk matrix for SIR model (\eqn{\Psi})
}
