% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mosquito_interface.R
\name{compute_q}
\alias{compute_q}
\title{Compute human blood feeding fraction (\eqn{q})}
\usage{
compute_q(model)
}
\arguments{
\item{model}{an object from \link{make_MicroMoB}}
}
\value{
a vector of length \code{p}
}
\description{
This method dispatches on the type of \code{model$mosquito}
}
