% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mosquito_RM.R
\name{compute_q.RM}
\alias{compute_q.RM}
\title{Compute human blood feeding fraction for RM model (\eqn{q})}
\usage{
\method{compute_q}{RM}(model)
}
\arguments{
\item{model}{an object from \link{make_MicroMoB}}
}
\description{
Compute human blood feeding fraction for RM model (\eqn{q})
}
