% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/time_spent.R
\name{compute_W}
\alias{compute_W}
\title{Compute human availability (W)}
\usage{
compute_W(human, Psi_t, t)
}
\arguments{
\item{human}{an object from \link{setup_human}}

\item{Psi_t}{either a matrix from \link{compute_Psi.day} or an array
from \link{compute_Psi.dt}}

\item{t}{time}
}
\description{
Compute human availability (W)
}
\seealso{
\link{compute_W.day} \link{compute_W.dt}
}
