% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/humans_SIR.R
\name{compute_W.SIR}
\alias{compute_W.SIR}
\title{Compute available humans for SIR model (\eqn{W})}
\usage{
\method{compute_W}{SIR}(model)
}
\arguments{
\item{model}{an object from \link{make_MicroMoB}}
}
\description{
Compute available humans for SIR model (\eqn{W})
}
