% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/time_spent.R
\name{compute_W.timespent.day}
\alias{compute_W.timespent.day}
\title{Compute human availability for daily time spent (W)}
\usage{
\method{compute_W.timespent}{day}(tisp, biteweight, human, t)
}
\arguments{
\item{tisp}{an object from \link{setup.timespent}}

\item{biteweight}{an object from \link{setup.biteweight}}

\item{human}{an object from \link{setup.human}}

\item{t}{time}
}
\description{
Compute human availability for daily time spent (W)
}
