% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/humans_interface.R
\name{compute_wf}
\alias{compute_wf}
\title{Compute human biting weights (\eqn{w_{f}})}
\usage{
compute_wf(model)
}
\arguments{
\item{model}{an object from \link{make_MicroMoB}}
}
\value{
a vector of length \code{n}
}
\description{
This method dispatches on the type of \code{model$human}.
}
