% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/humans_interface.R
\name{compute_x}
\alias{compute_x}
\title{Compute net infectiousness of humans (\eqn{x})}
\usage{
compute_x(model)
}
\arguments{
\item{model}{an object from \link{make_MicroMoB}}
}
\value{
a vector of length \code{n} giving the net infectiousness of human hosts in each stratum
}
\description{
In a Ross-Macdonald style transmission model, this is computed as
\deqn{x = c X} This method dispatches on the type of \code{model$human}.
}
