% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pathogen.R
\name{compute_x.rm}
\alias{compute_x.rm}
\title{Compute resident net infectiousness for Ross-Macdonland pathogen model (\eqn{x})}
\usage{
\method{compute_x}{rm}(model, t)
}
\arguments{
\item{model}{an object from \link{setup_model_object}}

\item{t}{time}
}
\value{
a vector of dimension \eqn{p \times 1}
}
\description{
Compute resident net infectiousness for Ross-Macdonland pathogen model (\eqn{x})
}
