% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/humans_SIS.R
\name{compute_x.SIS}
\alias{compute_x.SIS}
\title{Compute net infectiousness for SIR model}
\usage{
\method{compute_x}{SIS}(model)
}
\arguments{
\item{model}{an object from \link{make_MicroMoB}}
}
\description{
Compute net infectiousness for SIR model
}
