% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/other_hosts.R
\name{compute_x_delta}
\alias{compute_x_delta}
\title{Compute visitor prevalence (\eqn{x_{\delta}})}
\usage{
compute_x_delta(visitors, t)
}
\arguments{
\item{visitors}{an object from \link{setup_visitors}}

\item{t}{time}
}
\description{
Compute visitor prevalence (\eqn{x_{\delta}})
}
\seealso{
\link{compute_x_delta.simple}
}
