% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/visitor_trace.R
\name{compute_xd.trace}
\alias{compute_xd.trace}
\title{Compute net infectiousness of visitors for trace model (\eqn{x_{\delta}})}
\usage{
\method{compute_xd}{trace}(model)
}
\arguments{
\item{model}{an object from \link{make_MicroMoB}}
}
\description{
Compute net infectiousness of visitors for trace model (\eqn{x_{\delta}})
}
