% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mosquito.R
\name{compute_Z}
\alias{compute_Z}
\title{Compute infectious mosquitoes (\eqn{Z})}
\usage{
compute_Z(mosquito, t)
}
\arguments{
\item{mosquito}{an object from \link{setup_mosquito}}

\item{t}{time}
}
\description{
Compute infectious mosquitoes (\eqn{Z})
}
\seealso{
\link{compute_Z.trace}
}
