% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mosquito_trace.R
\name{compute_Z.trace}
\alias{compute_Z.trace}
\title{Compute null density of infective mosquitoes (\eqn{Z})}
\usage{
\method{compute_Z}{trace}(model)
}
\arguments{
\item{model}{an object from \link{make_MicroMoB}}
}
\value{
no return value
}
\description{
Compute null density of infective mosquitoes (\eqn{Z})
}
