% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/eip_fixed.R
\name{EIP.fixed}
\alias{EIP.fixed}
\title{Modify parameters due to exogenous forcing by all kinds of control}
\usage{
\method{EIP}{fixed}(t, MYZpar)
}
\arguments{
\item{t}{current simulation time}

\item{MYZpar}{a \link{list}}
}
\value{
the EIP maturation vector G \link{numeric}
}
\description{
Implements \link{EIP} for the fixed model (the EIP is constant)
}
