% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MicroMoB.R
\docType{package}
\name{MicroMoB}
\alias{MicroMoB}
\alias{MicroMoB-package}
\title{MicroMoB: Microsimulation for mosquito-borne pathogens}
\description{
Discrete time simulation of mosquito-borne pathogen transmission
}
\seealso{
Useful links:
\itemize{
  \item \url{https://dd-harp.github.io/MicroMoB/}
  \item \url{https://github.com/dd-harp/MicroMoB}
  \item Report bugs at \url{https://github.com/dd-harp/MicroMoB/issues}
}

}
\author{
\strong{Maintainer}: Sean L. Wu \email{slwood89@gmail.com} (\href{https://orcid.org/0000-0002-5781-9493}{ORCID})

Authors:
\itemize{
  \item David L. Smith \email{smitdave@uw.edu} (\href{https://orcid.org/0000-0003-4367-3849}{ORCID})
}

Other contributors:
\itemize{
  \item Sophie Libkind [contributor]
}

}
