% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Human.R
\name{setup.biteweight}
\alias{setup.biteweight}
\title{Setup a time spent model}
\usage{
setup.biteweight(type, model, ...)
}
\arguments{
\item{type}{a character in \code{c("null")}}

\item{model}{a model object (an \link{environment})}

\item{...}{other arguments to be passed to type methods}
}
\description{
Setup a time spent model. The model object must have already
been initialized with a human object (see \link{setup.human}). This adds
a \link{list} to \code{model} named "tisp" (Time spent).
}
