% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Human.R
\name{setup.human}
\alias{setup.human}
\alias{setup.human.strata}
\title{Setup a human object}
\usage{
setup.human(type, model, ...)

\method{setup.human}{strata}(type, model, H, J = NULL, ...)
}
\arguments{
\item{type}{a character in \code{c("simple", "strata")}}

\item{model}{a model object (an \link{environment})}

\item{...}{other arguments to be passed to type methods}

\item{H}{a vector of human population sizes}

\item{J}{a matrix whose columns assign human strata to patches (rows); the
columns must all sum to one.}
}
\description{
Setup a human object
}
