% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/human.R
\name{setup.human}
\alias{setup.human}
\title{Setup a human object}
\usage{
setup.human(type, model, ...)
}
\arguments{
\item{type}{a character in \code{c("strata")}}

\item{model}{a model object (an \link{environment})}

\item{...}{other arguments to be passed to type methods}
}
\description{
Setup a human object. The \code{model} object will have a list named \code{human}
added to it.
}
\section{strata}{
 see \link{setup.human.strata}
}

