% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Human.R
\name{setup.human}
\alias{setup.human}
\alias{setup.human.strata}
\title{Setup a human object}
\usage{
setup.human(type, model, ...)

\method{setup.human}{strata}(type, model, H, J = NULL, ...)
}
\arguments{
\item{type}{a character in \code{c("simple", "strata")}}

\item{model}{a model object (an \link{environment})}

\item{...}{other arguments to be passed to type methods}

\item{H}{a vector of human population sizes}

\item{J}{a matrix whose columns assign human strata to patches (rows); the
columns must all sum to one.}
}
\description{
Setup a human object. The \code{model} object will have a list named \code{human}
added to it. Within \code{human} the \code{H} vector is a vector of populations, in each strata,
in each patch. This means that if there are 2 strata and 3 patches, the first 3
elements give the distribution of people in strata 1 across the 3 patches.
}
