% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Human.R
\name{setup.timespent}
\alias{setup.timespent}
\alias{setup.timespent.null}
\title{Setup a time spent model}
\usage{
setup.timespent(type, model, ...)

\method{setup.timespent}{null}(type, model)
}
\arguments{
\item{type}{a character in \code{c("null", "matrix")}}

\item{model}{a model object (an \link{environment})}

\item{...}{other arguments to be passed to type methods}
}
\description{
Setup a time spent model. The model object must have already
been initialized with a human object (see \link{setup.human}).
}
\section{Null time spent model}{

The null model will simply assign each person to spend all time at their
home residence patch.
}

