% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Human.R
\name{setup.timespent}
\alias{setup.timespent}
\alias{setup.timespent.matrix_day}
\title{Setup a time spent model}
\usage{
setup.timespent(type, model, ...)

\method{setup.timespent}{matrix_day}(type, model, theta = NULL, ...)
}
\arguments{
\item{type}{a character in \code{c("matrix_day")}}

\item{model}{a model object (an \link{environment})}

\item{...}{other arguments to be passed to type methods}

\item{theta}{a time spent matrix, if \code{NULL} the identity matrix is used
(everyone stays at their home patch)}
}
\description{
Setup a time spent model. The model object must have already
been initialized with a human object (see \link{setup.human}). This adds
a \link{list} to \code{model} named "tisp" (Time spent).
}
\section{Matrix time spent model}{

When using \code{type = "matrix_day"} the model will assume that \code{theta} is a
matrix whose rows sum to \eqn{\le 1} giving the daily time spent distribution.
}

