% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/time_spent.R
\name{setup.timespent}
\alias{setup.timespent}
\title{Setup a time spent model}
\usage{
setup.timespent(type, model, ...)
}
\arguments{
\item{type}{a character in \code{c("day")}}

\item{model}{a model object (an \link{environment})}

\item{...}{other arguments to be passed to type methods}
}
\description{
Setup a time spent model. The model object must have already
been initialized with a human object (see \link{setup.human}). This adds
a \link{list} to \code{model$human} named "tisp" (Time spent).
}
\section{Daily time spent model}{
 see \link{setup.timespent.day}
}

\section{Fractional day time spent model}{
 see \link{setup.timespent.dt}
}

