% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/aquatic_trace.R
\name{setup_aqua_trace}
\alias{setup_aqua_trace}
\title{Setup aquatic (immature) mosquito model with trace (forced) emergence}
\usage{
setup_aqua_trace(model, lambda, stochastic)
}
\arguments{
\item{model}{an object from \link{make_MicroMoB}}

\item{lambda}{daily emergence of mosquitoes, may be time and patch varying, see \link{time_patch_varying_parameter}}

\item{stochastic}{should the model update deterministically or stochastically?}
}
\value{
no return value
}
\description{
Emergence is passed as a (possibly time varying) parameter which is
decoupled from the adult mosquito dynamics. This module assumes \code{l} and \code{p}
are equivalent, as emergence rates are given for \code{p}.
}
