% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/human.R
\name{setup_biteweight}
\alias{setup_biteweight}
\title{Setup a biting weight model}
\usage{
setup_biteweight(type, model, ...)
}
\arguments{
\item{type}{a character in \code{c("simple")}}

\item{model}{a model object (an \link{environment})}

\item{...}{other arguments to be passed to type methods}
}
\description{
Setup a biting weight model. The model object must have already
been initialized with a human object (see \link{setup_human}). This adds
a \link{list} to \code{model$human} named \code{biteweight} (Time spent).
}
\seealso{
\link{setup_biteweight.simple}
}
