% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/other_hosts.R
\name{setup_otherhosts.simple}
\alias{setup_otherhosts.simple}
\title{Setup simple other blood hosts}
\usage{
\method{setup_otherhosts}{simple}(type, model, B = NULL, zeta = 1, ...)
}
\arguments{
\item{type}{a character in \code{c("simple")}}

\item{model}{a model object (an \link{environment})}

\item{B}{an optional vector of search weights by patch, otherwise will
be set to all 0.}

\item{zeta}{search weights will be weighted as \eqn{B^{\Zeta}}}

\item{...}{other arguments to be passed to type methods}
}
\description{
Setup simple other blood hosts
}
