% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/other_hosts.R
\name{setup_otherhosts.simple}
\alias{setup_otherhosts.simple}
\title{Setup simple other blood hosts}
\usage{
\method{setup_otherhosts}{simple}(type, model, B = NULL, ...)
}
\arguments{
\item{type}{a character in \code{c("simple")}}

\item{model}{a model object (an \link{environment})}

\item{...}{other arguments to be passed to type methods}

\item{theta}{a time spent matrix, if \code{NULL} the identity matrix is used
(everyone stays at their home patch)}
}
\description{
Setup simple other blood hosts
}
