% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/humans_MOI.R
\name{step_humans.MOI}
\alias{step_humans.MOI}
\title{Update MOI human model}
\usage{
\method{step_humans}{MOI}(model)
}
\arguments{
\item{model}{an object from \link{make_MicroMoB}}
}
\description{
Update MOI human model
}
